refactor: move ApiDuplicatorService methods to a dedicated implementation
