+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/ApiDuplicatorServiceImpl.java
@@ -1,0 +1,674 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.rest.api.service.impl;
+
+import static java.util.Collections.singletonList;
+import static java.util.Objects.requireNonNull;
+import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toSet;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import io.gravitee.common.http.HttpMethod;
+import io.gravitee.definition.model.DefinitionVersion;
+import io.gravitee.definition.model.Proxy;
+import io.gravitee.definition.model.VirtualHost;
+import io.gravitee.rest.api.model.*;
+import io.gravitee.rest.api.model.api.ApiEntity;
+import io.gravitee.rest.api.model.api.DuplicateApiEntity;
+import io.gravitee.rest.api.model.api.UpdateApiEntity;
+import io.gravitee.rest.api.model.documentation.PageQuery;
+import io.gravitee.rest.api.model.permissions.RoleScope;
+import io.gravitee.rest.api.model.plan.PlanQuery;
+import io.gravitee.rest.api.service.*;
+import io.gravitee.rest.api.service.common.GraviteeContext;
+import io.gravitee.rest.api.service.exceptions.PlanNotFoundException;
+import io.gravitee.rest.api.service.exceptions.TechnicalManagementException;
+import io.gravitee.rest.api.service.exceptions.UserNotFoundException;
+import io.gravitee.rest.api.service.jackson.ser.api.ApiSerializer;
+import io.gravitee.rest.api.service.sanitizer.UrlSanitizerUtils;
+import io.gravitee.rest.api.service.spring.ImportConfiguration;
+import io.vertx.core.buffer.Buffer;
+import java.util.*;
+import java.util.stream.Collectors;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ApiDuplicatorServiceImpl extends AbstractApiService implements ApiDuplicatorService {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ApiDuplicatorServiceImpl.class);
+
+    @Autowired
+    private HttpClientService httpClientService;
+
+    @Autowired
+    private ImportConfiguration importConfiguration;
+
+    @Autowired
+    private MediaService mediaService;
+
+    @Override
+    public ApiEntity createWithImportedDefinition(ApiEntity apiEntity, String apiDefinitionOrURL, String userId) {
+        String apiDefinition = fetchApiDefinitionContentFromURL(apiDefinitionOrURL);
+        try {
+            // Read the whole definition
+            final JsonNode jsonNode = objectMapper.readTree(apiDefinition);
+            UpdateApiEntity importedApi = convertToEntity(apiDefinition, jsonNode);
+            ApiEntity createdApiEntity = this.create0(importedApi, userId, false, jsonNode, GraviteeContext.getCurrentEnvironment());
+            createPageAndMedia(createdApiEntity, jsonNode);
+            updateApiReferences(createdApiEntity, jsonNode);
+            return createdApiEntity;
+        } catch (JsonProcessingException e) {
+            LOGGER.error("An error occurs while trying to JSON deserialize the API {}", apiDefinition, e);
+            throw new TechnicalManagementException("An error occurs while trying to JSON deserialize the API definition.");
+        }
+    }
+
+    @Override
+    public ApiEntity duplicate(final ApiEntity apiEntity, final DuplicateApiEntity duplicateApiEntity) {
+        requireNonNull(apiEntity, "Missing ApiEntity");
+        final String apiId = apiEntity.getId();
+        LOGGER.debug("Duplicate API {}", apiId);
+
+        final UpdateApiEntity newApiEntity = convert(apiEntity);
+        final Proxy proxy = apiEntity.getProxy();
+        proxy.setVirtualHosts(singletonList(new VirtualHost(duplicateApiEntity.getContextPath())));
+        newApiEntity.setProxy(proxy);
+        newApiEntity.setVersion(duplicateApiEntity.getVersion() == null ? apiEntity.getVersion() : duplicateApiEntity.getVersion());
+
+        if (duplicateApiEntity.getFilteredFields().contains("groups")) {
+            newApiEntity.setGroups(null);
+        } else {
+            newApiEntity.setGroups(apiEntity.getGroups());
+        }
+        final ApiEntity duplicatedApi =
+            this.create0(newApiEntity, getAuthenticatedUsername(), false, null, GraviteeContext.getCurrentEnvironment());
+
+        if (!duplicateApiEntity.getFilteredFields().contains("members")) {
+            final Set<MembershipEntity> membershipsToDuplicate = membershipService.getMembershipsByReference(
+                io.gravitee.rest.api.model.MembershipReferenceType.API,
+                apiId
+            );
+            RoleEntity primaryOwnerRole = roleService.findPrimaryOwnerRoleByOrganization(
+                GraviteeContext.getCurrentOrganization(),
+                RoleScope.API
+            );
+            if (primaryOwnerRole != null) {
+                String primaryOwnerRoleId = primaryOwnerRole.getId();
+                membershipsToDuplicate.forEach(
+                    membership -> {
+                        String roleId = membership.getRoleId();
+                        if (!primaryOwnerRoleId.equals(roleId)) {
+                            membershipService.addRoleToMemberOnReference(
+                                io.gravitee.rest.api.model.MembershipReferenceType.API,
+                                duplicatedApi.getId(),
+                                membership.getMemberType(),
+                                membership.getMemberId(),
+                                roleId
+                            );
+                        }
+                    }
+                );
+            }
+        }
+
+        if (!duplicateApiEntity.getFilteredFields().contains("pages")) {
+            final List<PageEntity> pages = pageService.search(new PageQuery.Builder().api(apiId).build(), true);
+            pages.forEach(page -> pageService.create(duplicatedApi.getId(), page));
+        }
+
+        if (!duplicateApiEntity.getFilteredFields().contains("plans")) {
+            final Set<PlanEntity> plans = planService.findByApi(apiId);
+            plans.forEach(plan -> planService.create(duplicatedApi.getId(), plan));
+        }
+
+        return duplicatedApi;
+    }
+
+    @Override
+    public String exportAsJson(final String apiId, String exportVersion, String... filteredFields) {
+        ApiEntity apiEntity = this.findById(apiId);
+        // set metadata for serialize process
+        Map<String, Object> metadata = new HashMap<>();
+        metadata.put(ApiSerializer.METADATA_EXPORT_VERSION, exportVersion);
+        metadata.put(ApiSerializer.METADATA_FILTERED_FIELDS_LIST, Arrays.asList(filteredFields));
+        apiEntity.setMetadata(metadata);
+
+        try {
+            return objectMapper.writeValueAsString(apiEntity);
+        } catch (final Exception e) {
+            LOGGER.error("An error occurs while trying to JSON serialize the API {}", apiEntity, e);
+        }
+        return "";
+    }
+
+    @Override
+    public ApiEntity updateWithImportedDefinition(ApiEntity apiEntity, String apiDefinitionOrURL, String userId) {
+        String apiDefinition = fetchApiDefinitionContentFromURL(apiDefinitionOrURL);
+        try {
+            // Read the whole definition
+            final JsonNode jsonNode = objectMapper.readTree(apiDefinition);
+            UpdateApiEntity importedApi = convertToEntity(apiDefinition, jsonNode);
+            ApiEntity updatedApiEntity = this.update(apiEntity.getId(), importedApi, false);
+            updateApiReferences(updatedApiEntity, jsonNode);
+            return updatedApiEntity;
+        } catch (JsonProcessingException e) {
+            LOGGER.error("An error occurs while trying to JSON deserialize the API {}", apiDefinition, e);
+            throw new TechnicalManagementException("An error occurs while trying to JSON deserialize the API definition.");
+        }
+    }
+
+    private UpdateApiEntity convertToEntity(String apiDefinition, JsonNode jsonNode) throws JsonProcessingException {
+        final UpdateApiEntity importedApi = objectMapper
+            // because definition could contains other values than the api itself (pages, members)
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+            .readValue(apiDefinition, UpdateApiEntity.class);
+
+        // Initialize with a default path
+        if (
+            Objects.equals(importedApi.getGraviteeDefinitionVersion(), DefinitionVersion.V1.getLabel()) &&
+            (importedApi.getPaths() == null || importedApi.getPaths().isEmpty())
+        ) {
+            importedApi.setPaths(Collections.singletonMap("/", new ArrayList<>()));
+        }
+
+        //create group if not exist & replace groupName by groupId
+        if (importedApi.getGroups() != null) {
+            Set<String> groupNames = new HashSet<>(importedApi.getGroups());
+            importedApi.getGroups().clear();
+            for (String name : groupNames) {
+                List<GroupEntity> groupEntities = groupService.findByName(name);
+                GroupEntity group;
+                if (groupEntities.isEmpty()) {
+                    NewGroupEntity newGroupEntity = new NewGroupEntity();
+                    newGroupEntity.setName(name);
+                    group = groupService.create(newGroupEntity);
+                } else {
+                    group = groupEntities.get(0);
+                }
+                importedApi.getGroups().add(group.getId());
+            }
+        }
+
+        // Views & Categories
+        // Before 3.0.2, API 'categories' were called 'views'. This is for compatibility.
+        final JsonNode viewsDefinition = jsonNode.path("views");
+        if (viewsDefinition != null && viewsDefinition.isArray()) {
+            Set<String> categories = new HashSet<>();
+            for (JsonNode viewNode : viewsDefinition) {
+                categories.add(viewNode.asText());
+            }
+            importedApi.setCategories(categories);
+        }
+
+        return importedApi;
+    }
+
+    private void createPageAndMedia(ApiEntity createdApiEntity, JsonNode jsonNode) {
+        final JsonNode apiMedia = jsonNode.path("apiMedia");
+        if (apiMedia != null && apiMedia.isArray()) {
+            for (JsonNode media : apiMedia) {
+                mediaService.createWithDefinition(createdApiEntity.getId(), media.toString());
+            }
+        }
+
+        final JsonNode pages = jsonNode.path("pages");
+        if (pages != null && pages.isArray()) {
+            for (JsonNode page : pages) {
+                PageEntity pageEntity = pageService.createWithDefinition(createdApiEntity.getId(), page.toString());
+                ((ObjectNode) page).put("id", pageEntity.getId());
+            }
+        }
+
+        List<PageEntity> search = pageService.search(
+            new PageQuery.Builder()
+                .api(createdApiEntity.getId())
+                .name(SystemFolderType.ASIDE.folderName())
+                .type(PageType.SYSTEM_FOLDER)
+                .build()
+        );
+        if (search.isEmpty()) {
+            this.createSystemFolder(createdApiEntity.getId());
+        }
+    }
+
+    private void createOrUpdateChildrenPages(String apiId, String parentId, List<PageEntityTreeNode> children) {
+        for (final PageEntityTreeNode child : children) {
+            PageEntity pageEntityToImport = child.data;
+            pageEntityToImport.setParentId(parentId);
+
+            PageEntity createdOrUpdatedPage = null;
+            if (pageEntityToImport.getId() != null) {
+                createdOrUpdatedPage = pageService.findById(pageEntityToImport.getId());
+            } else {
+                PageQuery query = new PageQuery.Builder()
+                    .api(apiId)
+                    .name(pageEntityToImport.getName())
+                    .type(PageType.valueOf(pageEntityToImport.getType()))
+                    .build();
+
+                List<PageEntity> pages = pageService.search(query);
+                if (pages.size() == 1) {
+                    createdOrUpdatedPage = pages.get(0);
+                } else if (pages.size() > 1) {
+                    LOGGER.error(
+                        "Not able to identify the page to update: {}. Too much pages with the same name",
+                        pageEntityToImport.getName()
+                    );
+                    throw new TechnicalManagementException(
+                        "Not able to identify the page to update: " + pageEntityToImport.getName() + ". Too much pages with the same name"
+                    );
+                }
+            }
+
+            if (createdOrUpdatedPage == null) {
+                NewPageEntity newPage = new NewPageEntity();
+                newPage.setConfiguration(pageEntityToImport.getConfiguration());
+                newPage.setContent(pageEntityToImport.getContent());
+                newPage.setExcludedAccessControls(pageEntityToImport.isExcludedAccessControls());
+                newPage.setAccessControls(pageEntityToImport.getAccessControls());
+                newPage.setHomepage(pageEntityToImport.isHomepage());
+                newPage.setLastContributor(pageEntityToImport.getLastContributor());
+                newPage.setName(pageEntityToImport.getName());
+                newPage.setOrder(pageEntityToImport.getOrder());
+                newPage.setParentId(pageEntityToImport.getParentId());
+                newPage.setPublished(pageEntityToImport.isPublished());
+                newPage.setSource(pageEntityToImport.getSource());
+                newPage.setType(PageType.valueOf(pageEntityToImport.getType()));
+                newPage.setAttachedMedia(pageEntityToImport.getAttachedMedia());
+                createdOrUpdatedPage = pageService.createPage(apiId, newPage);
+            } else {
+                UpdatePageEntity updatePageEntity = new UpdatePageEntity();
+                updatePageEntity.setConfiguration(pageEntityToImport.getConfiguration());
+                updatePageEntity.setContent(pageEntityToImport.getContent());
+                updatePageEntity.setExcludedAccessControls(pageEntityToImport.isExcludedAccessControls());
+                updatePageEntity.setAccessControls(pageEntityToImport.getAccessControls());
+                updatePageEntity.setHomepage(pageEntityToImport.isHomepage());
+                updatePageEntity.setLastContributor(pageEntityToImport.getLastContributor());
+                updatePageEntity.setName(pageEntityToImport.getName());
+                updatePageEntity.setOrder(pageEntityToImport.getOrder());
+                updatePageEntity.setParentId(pageEntityToImport.getParentId());
+                updatePageEntity.setPublished(pageEntityToImport.isPublished());
+                updatePageEntity.setSource(pageEntityToImport.getSource());
+                updatePageEntity.setAttachedMedia(pageEntityToImport.getAttachedMedia());
+
+                createdOrUpdatedPage = pageService.update(createdOrUpdatedPage.getId(), updatePageEntity);
+            }
+
+            if (child.children != null && !child.children.isEmpty()) {
+                this.createOrUpdateChildrenPages(apiId, createdOrUpdatedPage.getId(), child.children);
+            }
+        }
+    }
+
+    private String fetchApiDefinitionContentFromURL(String apiDefinitionOrURL) {
+        if (apiDefinitionOrURL.toUpperCase().startsWith("HTTP")) {
+            UrlSanitizerUtils.checkAllowed(
+                apiDefinitionOrURL,
+                importConfiguration.getImportWhitelist(),
+                importConfiguration.isAllowImportFromPrivate()
+            );
+            Buffer buffer = httpClientService.request(HttpMethod.GET, apiDefinitionOrURL, null, null, null);
+            return buffer.toString();
+        }
+        return apiDefinitionOrURL;
+    }
+
+    private void updateApiReferences(ApiEntity createdOrUpdatedApiEntity, JsonNode jsonNode) throws JsonProcessingException {
+        // Members
+        final JsonNode membersToImport = jsonNode.path("members");
+        if (membersToImport != null && membersToImport.isArray()) {
+            // get current members of the api
+            Set<MemberToImport> membersAlreadyPresent = membershipService
+                .getMembersByReference(MembershipReferenceType.API, createdOrUpdatedApiEntity.getId())
+                .stream()
+                .map(
+                    member -> {
+                        UserEntity userEntity = userService.findById(member.getId());
+                        return new MemberToImport(
+                            userEntity.getSource(),
+                            userEntity.getSourceId(),
+                            member.getRoles().stream().map(RoleEntity::getId).collect(Collectors.toList()),
+                            null
+                        );
+                    }
+                )
+                .collect(toSet());
+            // get the current PO
+            RoleEntity poRole = roleService.findPrimaryOwnerRoleByOrganization(GraviteeContext.getCurrentOrganization(), RoleScope.API);
+            if (poRole != null) {
+                String poRoleId = poRole.getId();
+                MemberToImport currentPo = membersAlreadyPresent
+                    .stream()
+                    .filter(memberToImport -> memberToImport.getRoles().contains(poRoleId))
+                    .findFirst()
+                    .orElse(new MemberToImport());
+
+                List<String> roleUsedInTransfert = null;
+                MemberToImport futurePO = null;
+
+                // upsert members
+                for (final JsonNode memberNode : membersToImport) {
+                    MemberToImport memberToImport = objectMapper.readValue(memberNode.toString(), MemberToImport.class);
+                    String roleToAdd = memberToImport.getRole();
+                    List<String> rolesToImport = memberToImport.getRoles();
+                    if (roleToAdd != null && !roleToAdd.isEmpty()) {
+                        if (rolesToImport == null) {
+                            rolesToImport = new ArrayList<>();
+                            memberToImport.setRoles(rolesToImport);
+                        }
+                        Optional<RoleEntity> optRoleToAddEntity = roleService.findByScopeAndName(RoleScope.API, roleToAdd);
+                        if (optRoleToAddEntity.isPresent()) {
+                            rolesToImport.add(optRoleToAddEntity.get().getId());
+                        } else {
+                            LOGGER.warn("Role {} does not exist", roleToAdd);
+                        }
+                    }
+                    if (rolesToImport != null) {
+                        rolesToImport.sort(Comparator.naturalOrder());
+                    }
+                    boolean presentWithSameRole =
+                        memberToImport.getRoles() != null &&
+                        !memberToImport.getRoles().isEmpty() &&
+                        membersAlreadyPresent
+                            .stream()
+                            .anyMatch(
+                                m -> {
+                                    m.getRoles().sort(Comparator.naturalOrder());
+                                    return (
+                                        m.getRoles().equals(memberToImport.getRoles()) &&
+                                        (
+                                            m.getSourceId().equals(memberToImport.getSourceId()) &&
+                                            m.getSource().equals(memberToImport.getSource())
+                                        )
+                                    );
+                                }
+                            );
+
+                    // add/update members if :
+                    //  - not already present with the same role
+                    //  - not the new PO
+                    //  - not the current PO
+                    if (
+                        !presentWithSameRole &&
+                        (
+                            memberToImport.getRoles() != null &&
+                            !memberToImport.getRoles().isEmpty() &&
+                            !memberToImport.getRoles().contains(poRoleId)
+                        ) &&
+                        !(
+                            memberToImport.getSourceId().equals(currentPo.getSourceId()) &&
+                            memberToImport.getSource().equals(currentPo.getSource())
+                        )
+                    ) {
+                        try {
+                            UserEntity userEntity = userService.findBySource(
+                                memberToImport.getSource(),
+                                memberToImport.getSourceId(),
+                                false
+                            );
+
+                            rolesToImport.forEach(
+                                role ->
+                                    membershipService.addRoleToMemberOnReference(
+                                        MembershipReferenceType.API,
+                                        createdOrUpdatedApiEntity.getId(),
+                                        MembershipMemberType.USER,
+                                        userEntity.getId(),
+                                        role
+                                    )
+                            );
+                        } catch (UserNotFoundException unfe) {}
+                    }
+
+                    // get the future role of the current PO
+                    if (
+                        currentPo.getSourceId().equals(memberToImport.getSourceId()) &&
+                        currentPo.getSource().equals(memberToImport.getSource()) &&
+                        !rolesToImport.contains(poRoleId)
+                    ) {
+                        roleUsedInTransfert = rolesToImport;
+                    }
+
+                    if (rolesToImport.contains(poRoleId)) {
+                        futurePO = memberToImport;
+                    }
+                }
+
+                // transfer the ownership
+                if (
+                    futurePO != null &&
+                    !(currentPo.getSource().equals(futurePO.getSource()) && currentPo.getSourceId().equals(futurePO.getSourceId()))
+                ) {
+                    try {
+                        UserEntity userEntity = userService.findBySource(futurePO.getSource(), futurePO.getSourceId(), false);
+                        List<RoleEntity> roleEntity = null;
+                        if (roleUsedInTransfert != null && !roleUsedInTransfert.isEmpty()) {
+                            roleEntity = roleUsedInTransfert.stream().map(roleService::findById).collect(Collectors.toList());
+                        }
+                        membershipService.transferApiOwnership(
+                            createdOrUpdatedApiEntity.getId(),
+                            new MembershipService.MembershipMember(userEntity.getId(), null, MembershipMemberType.USER),
+                            roleEntity
+                        );
+                    } catch (UserNotFoundException unfe) {}
+                }
+            }
+        }
+
+        //Pages
+        final JsonNode pagesDefinition = jsonNode.path("pages");
+        if (pagesDefinition != null && pagesDefinition.isArray()) {
+            List<PageEntity> pagesList = objectMapper.readValue(
+                pagesDefinition.toString(),
+                objectMapper.getTypeFactory().constructCollectionType(List.class, PageEntity.class)
+            );
+            PageEntityTreeNode documentationTree = new PageEntityTreeNode(new PageEntity());
+            documentationTree.appendListToTree(pagesList);
+            createOrUpdateChildrenPages(createdOrUpdatedApiEntity.getId(), null, documentationTree.children);
+        }
+
+        //Plans
+        final JsonNode plansDefinition = jsonNode.path("plans");
+        if (plansDefinition != null && plansDefinition.isArray()) {
+            for (JsonNode planNode : plansDefinition) {
+                if (planNode.has("id")) {
+                    try {
+                        PlanEntity plan = planService.findById(planNode.get("id").asText());
+                        UpdatePlanEntity updatePlanEntity = objectMapper.readValue(planNode.toString(), UpdatePlanEntity.class);
+                        updatePlanEntity.setId(plan.getId());
+                        planService.update(updatePlanEntity);
+                    } catch (PlanNotFoundException npe) {
+                        NewPlanEntity newPlanEntity = objectMapper.readValue(planNode.toString(), NewPlanEntity.class);
+                        newPlanEntity.setApi(createdOrUpdatedApiEntity.getId());
+                        planService.create(newPlanEntity);
+                    }
+                } else {
+                    PlanQuery query = new PlanQuery.Builder()
+                        .api(createdOrUpdatedApiEntity.getId())
+                        .name(planNode.get("name").asText())
+                        .security(PlanSecurityType.valueOf(planNode.get("security").asText().toUpperCase()))
+                        .build();
+                    List<PlanEntity> planEntities = planService
+                        .search(query)
+                        .stream()
+                        .filter(planEntity -> !PlanStatus.CLOSED.equals(planEntity.getStatus()))
+                        .collect(toList());
+                    if (planEntities.isEmpty()) {
+                        NewPlanEntity newPlanEntity = objectMapper.readValue(planNode.toString(), NewPlanEntity.class);
+                        newPlanEntity.setApi(createdOrUpdatedApiEntity.getId());
+                        planService.create(newPlanEntity);
+                    } else if (planEntities.size() == 1) {
+                        UpdatePlanEntity updatePlanEntity = objectMapper.readValue(planNode.toString(), UpdatePlanEntity.class);
+                        updatePlanEntity.setId(planEntities.iterator().next().getId());
+                        planService.update(updatePlanEntity);
+                    } else {
+                        LOGGER.error(
+                            "Not able to identify the plan to update: {}. Too much plan with the same name",
+                            planNode.get("name").asText()
+                        );
+                        throw new TechnicalManagementException(
+                            "Not able to identify the plan to update: " +
+                            planNode.get("name").asText() +
+                            ". Too much plan with the same name"
+                        );
+                    }
+                }
+            }
+        }
+        // Metadata
+        final JsonNode metadataDefinition = jsonNode.path("metadata");
+        if (metadataDefinition != null && metadataDefinition.isArray()) {
+            try {
+                for (JsonNode metadataNode : metadataDefinition) {
+                    UpdateApiMetadataEntity updateApiMetadataEntity = objectMapper.readValue(
+                        metadataNode.toString(),
+                        UpdateApiMetadataEntity.class
+                    );
+                    updateApiMetadataEntity.setApiId(createdOrUpdatedApiEntity.getId());
+                    apiMetadataService.update(updateApiMetadataEntity);
+                }
+            } catch (Exception ex) {
+                LOGGER.error("An error occurs while creating API metadata", ex);
+                throw new TechnicalManagementException("An error occurs while creating API Metadata", ex);
+            }
+        }
+    }
+
+    private static class MemberToImport {
+
+        private String source;
+
+        private String sourceId;
+        private List<String> roles; // After v3
+        private String role; // Before v3
+
+        public MemberToImport() {}
+
+        public MemberToImport(String source, String sourceId, List<String> roles, String role) {
+            this.source = source;
+            this.sourceId = sourceId;
+            this.roles = roles;
+            this.role = role;
+        }
+
+        public String getSource() {
+            return source;
+        }
+
+        public void setSource(String source) {
+            this.source = source;
+        }
+
+        public String getSourceId() {
+            return sourceId;
+        }
+
+        public void setSourceId(String sourceId) {
+            this.sourceId = sourceId;
+        }
+
+        public List<String> getRoles() {
+            return roles;
+        }
+
+        public void setRoles(List<String> roles) {
+            this.roles = roles;
+        }
+
+        public String getRole() {
+            return role;
+        }
+
+        public void setRole(String role) {
+            this.role = role;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            MemberToImport that = (MemberToImport) o;
+
+            return source.equals(that.source) && sourceId.equals(that.sourceId);
+        }
+
+        @Override
+        public int hashCode() {
+            int result = source.hashCode();
+            result = 31 * result + sourceId.hashCode();
+            return result;
+        }
+    }
+
+    private static class PageEntityTreeNode {
+
+        PageEntity data;
+        PageEntityTreeNode parent;
+
+        List<PageEntityTreeNode> children;
+
+        public PageEntityTreeNode(PageEntity data) {
+            this.data = data;
+            this.children = new LinkedList<>();
+        }
+
+        public PageEntityTreeNode addChild(PageEntity child) {
+            PageEntityTreeNode childNode = new PageEntityTreeNode(child);
+            childNode.parent = this;
+            this.children.add(childNode);
+            return childNode;
+        }
+
+        private PageEntityTreeNode findById(String id) {
+            if (id.equals(data.getId())) {
+                return this;
+            }
+            for (PageEntityTreeNode child : children) {
+                PageEntityTreeNode result = child.findById(id);
+                if (result != null) {
+                    return result;
+                }
+            }
+            return null;
+        }
+
+        public void appendListToTree(List<PageEntity> pagesList) {
+            List<PageEntity> orphans = new ArrayList<>();
+            for (PageEntity newPage : pagesList) {
+                if (newPage.getParentId() == null || newPage.getParentId().isEmpty()) {
+                    this.addChild(newPage);
+                } else {
+                    PageEntityTreeNode parentNode = this.findById(newPage.getParentId());
+                    if (parentNode != null) {
+                        parentNode.addChild(newPage);
+                    } else {
+                        orphans.add(newPage);
+                    }
+                }
+            }
+            if (!orphans.isEmpty() && orphans.size() < pagesList.size()) {
+                appendListToTree(orphans);
+            }
+        }
+    }
+}
