+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/ApiServiceImpl.java
@@ -22,11 +22,9 @@ import static io.gravitee.rest.api.model.EventType.PUBLISH_API;
 import static io.gravitee.rest.api.model.ImportSwaggerDescriptorEntity.Type.INLINE;
 import static io.gravitee.rest.api.model.PageType.SWAGGER;
 import static io.gravitee.rest.api.model.WorkflowReferenceType.API;
-import static io.gravitee.rest.api.model.WorkflowState.DRAFT;
 import static io.gravitee.rest.api.model.WorkflowType.REVIEW;
 import static java.util.Collections.*;
 import static java.util.Comparator.comparing;
-import static java.util.Objects.requireNonNull;
 import static java.util.Optional.of;
 import static java.util.stream.Collectors.*;
 import static org.apache.commons.lang3.StringUtils.isBlank;
@@ -34,22 +32,15 @@ import static org.apache.commons.lang3.StringUtils.isBlank;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.base.Strings;
 import io.gravitee.common.component.Lifecycle;
 import io.gravitee.common.data.domain.Page;
-import io.gravitee.common.http.HttpMethod;
 import io.gravitee.definition.model.*;
-import io.gravitee.definition.model.Plan;
 import io.gravitee.definition.model.Properties;
 import io.gravitee.definition.model.endpoint.HttpEndpoint;
-import io.gravitee.definition.model.flow.Step;
-import io.gravitee.definition.model.services.discovery.EndpointDiscoveryService;
 import io.gravitee.definition.model.services.healthcheck.HealthCheckService;
 import io.gravitee.repository.exceptions.TechnicalException;
 import io.gravitee.repository.management.api.ApiQualityRuleRepository;
-import io.gravitee.repository.management.api.ApiRepository;
 import io.gravitee.repository.management.api.search.ApiCriteria;
 import io.gravitee.repository.management.api.search.ApiFieldExclusionFilter;
 import io.gravitee.repository.management.model.*;
@@ -57,12 +48,9 @@ import io.gravitee.repository.management.model.Api;
 import io.gravitee.repository.management.model.ApiLifecycleState;
 import io.gravitee.repository.management.model.Visibility;
 import io.gravitee.rest.api.model.*;
-import io.gravitee.rest.api.model.EntrypointReferenceType;
 import io.gravitee.rest.api.model.EventType;
 import io.gravitee.rest.api.model.MembershipMemberType;
 import io.gravitee.rest.api.model.MembershipReferenceType;
-import io.gravitee.rest.api.model.MetadataFormat;
-import io.gravitee.rest.api.model.TagReferenceType;
 import io.gravitee.rest.api.model.alert.AlertReferenceType;
 import io.gravitee.rest.api.model.alert.AlertTriggerEntity;
 import io.gravitee.rest.api.model.api.*;
@@ -72,57 +60,37 @@ import io.gravitee.rest.api.model.common.Pageable;
 import io.gravitee.rest.api.model.common.PageableImpl;
 import io.gravitee.rest.api.model.common.Sortable;
 import io.gravitee.rest.api.model.documentation.PageQuery;
-import io.gravitee.rest.api.model.notification.GenericNotificationConfigEntity;
 import io.gravitee.rest.api.model.parameters.Key;
 import io.gravitee.rest.api.model.parameters.ParameterReferenceType;
 import io.gravitee.rest.api.model.permissions.ApiPermission;
 import io.gravitee.rest.api.model.permissions.RolePermissionAction;
 import io.gravitee.rest.api.model.permissions.RoleScope;
-import io.gravitee.rest.api.model.permissions.SystemRole;
-import io.gravitee.rest.api.model.plan.PlanQuery;
-import io.gravitee.rest.api.model.settings.ApiPrimaryOwnerMode;
 import io.gravitee.rest.api.model.subscription.SubscriptionQuery;
 import io.gravitee.rest.api.service.*;
 import io.gravitee.rest.api.service.builder.EmailNotificationBuilder;
 import io.gravitee.rest.api.service.common.GraviteeContext;
-import io.gravitee.rest.api.service.common.RandomString;
 import io.gravitee.rest.api.service.exceptions.*;
 import io.gravitee.rest.api.service.impl.search.SearchResult;
-import io.gravitee.rest.api.service.impl.upgrade.DefaultMetadataUpgrader;
-import io.gravitee.rest.api.service.jackson.ser.api.ApiSerializer;
 import io.gravitee.rest.api.service.migration.APIV1toAPIV2Converter;
 import io.gravitee.rest.api.service.notification.ApiHook;
-import io.gravitee.rest.api.service.notification.HookScope;
 import io.gravitee.rest.api.service.notification.NotificationParamsBuilder;
-import io.gravitee.rest.api.service.notification.NotificationTemplateService;
 import io.gravitee.rest.api.service.processor.ApiSynchronizationProcessor;
-import io.gravitee.rest.api.service.sanitizer.UrlSanitizerUtils;
-import io.gravitee.rest.api.service.search.SearchEngineService;
 import io.gravitee.rest.api.service.search.query.Query;
 import io.gravitee.rest.api.service.search.query.QueryBuilder;
-import io.gravitee.rest.api.service.spring.ImportConfiguration;
-import io.vertx.core.buffer.Buffer;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.StringReader;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.*;
 import java.util.function.Predicate;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.regex.PatternSyntaxException;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.xml.bind.DatatypeConverter;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
 
@@ -133,114 +101,39 @@ import org.springframework.util.CollectionUtils;
  * @author GraviteeSource Team
  */
 @Component
-public class ApiServiceImpl extends AbstractService implements ApiService, ApiDuplicatorService {
+public class ApiServiceImpl extends AbstractApiService implements ApiService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ApiServiceImpl.class);
-
-    private static final Pattern DUPLICATE_SLASH_REMOVER = Pattern.compile("(?<!(http:|https:))[//]+");
-    // RFC 6454 section-7.1, serialized-origin regex from RFC 3986
-    private static final Pattern CORS_REGEX_PATTERN = Pattern.compile("^((\\*)|(null)|(^(([^:\\/?#]+):)?(\\/\\/([^\\/?#]*))?))$");
-    private static final String[] CORS_REGEX_CHARS = new String[] { "{", "[", "(", "*" };
-    private static final String URI_PATH_SEPARATOR = "/";
-
-    @Autowired
-    private ApiRepository apiRepository;
 
     @Autowired
     private ApiQualityRuleRepository apiQualityRuleRepository;
 
     @Autowired
-    private ObjectMapper objectMapper;
-
-    @Autowired
     private EventService eventService;
-
-    @Autowired
-    private UserService userService;
-
-    @Autowired
-    private PageService pageService;
-
-    @Autowired
-    private MembershipService membershipService;
-
-    @Autowired
-    private GroupService groupService;
-
-    @Autowired
-    private PlanService planService;
 
     @Autowired
     private ApiSynchronizationProcessor apiSynchronizationProcessor;
 
     @Autowired
-    private ApiMetadataService apiMetadataService;
-
-    @Autowired
     private SubscriptionService subscriptionService;
-
-    @Autowired
-    private AuditService auditService;
 
     @Autowired
     private TopApiService topApiService;
 
     @Autowired
-    private GenericNotificationConfigService genericNotificationConfigService;
-
-    @Autowired
     private PortalNotificationConfigService portalNotificationConfigService;
-
-    @Autowired
-    private NotifierService notifierService;
 
     @Autowired
     private SwaggerService swaggerService;
 
     @Autowired
-    private SearchEngineService searchEngineService;
-
-    @Autowired
     private ApiHeaderService apiHeaderService;
-
-    @Autowired
-    private NotificationTemplateService notificationTemplateService;
-
-    @Autowired
-    private ParameterService parameterService;
-
-    @Autowired
-    private TagService tagService;
-
-    @Autowired
-    private EntrypointService entrypointService;
-
-    @Autowired
-    private WorkflowService workflowService;
-
-    @Autowired
-    private HttpClientService httpClientService;
-
-    @Autowired
-    private VirtualHostService virtualHostService;
 
     @Autowired
     private AlertService alertService;
 
     @Autowired
-    private RoleService roleService;
-
-    @Autowired
-    private CategoryService categoryService;
-
-    @Autowired
     private ApplicationService applicationService;
-
-    @Autowired
-    private ImportConfiguration importConfiguration;
-
-    @Autowired
-    private PolicyService policyService;
 
     @Autowired
     private MediaService mediaService;
@@ -254,11 +147,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
     @Value("${configuration.default-api-icon:}")
     private String defaultApiIcon;
 
-    private static final Pattern LOGGING_MAX_DURATION_PATTERN = Pattern.compile(
-        "(?<before>.*)\\#request.timestamp\\s*\\<\\=?\\s*(?<timestamp>\\d*)l(?<after>.*)"
-    );
-    private static final String LOGGING_MAX_DURATION_CONDITION = "#request.timestamp <= %dl";
-    private static final String LOGGING_DELIMITER_BASE = "\\s+(\\|\\||\\&\\&)\\s+";
     private static final String ENDPOINTS_DELIMITER = "\n";
 
     @Override
@@ -272,38 +160,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
         createMetadata(swaggerApiEntity.getMetadata(), createdApi.getId());
 
         return createdApi;
-    }
-
-    private void checkGroupExistence(Set<String> groups) {
-        // check the existence of groups
-        if (groups != null && !groups.isEmpty()) {
-            try {
-                groupService.findByIds(new HashSet(groups));
-            } catch (GroupsNotFoundException gnfe) {
-                throw new InvalidDataException("These groups [" + gnfe.getParameters().get("groups") + "] do not exist");
-            }
-        }
-    }
-
-    private Set<String> removePOGroups(Set<String> groups, String apiId) {
-        Stream<GroupEntity> groupEntityStream = groupService.findByIds(groups).stream();
-
-        if (apiId != null) {
-            final MembershipEntity primaryOwner = membershipService.getPrimaryOwner(MembershipReferenceType.API, apiId);
-            if (primaryOwner.getMemberType() == MembershipMemberType.GROUP) {
-                // don't remove the primary owner group of this API.
-                groupEntityStream =
-                    groupEntityStream.filter(
-                        group -> StringUtils.isEmpty(group.getApiPrimaryOwner()) || group.getId().equals(primaryOwner.getMemberId())
-                    );
-            } else {
-                groupEntityStream = groupEntityStream.filter(group -> StringUtils.isEmpty(group.getApiPrimaryOwner()));
-            }
-        } else {
-            groupEntityStream = groupEntityStream.filter(group -> StringUtils.isEmpty(group.getApiPrimaryOwner()));
-        }
-
-        return groupEntityStream.map(GroupEntity::getId).collect(Collectors.toSet());
     }
 
     private void createMetadata(List<ApiMetadataEntity> apiMetadata, String apiId) {
@@ -378,7 +234,7 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
         final String userId,
         final ImportSwaggerDescriptorEntity swaggerDescriptor
     ) {
-        final ApiEntity createdApi = create0(apiEntity, userId);
+        final ApiEntity createdApi = this.create0(apiEntity, userId, true, null, GraviteeContext.getCurrentEnvironment());
         createOrUpdateDocumentation(swaggerDescriptor, createdApi, true);
         return createdApi;
     }
@@ -427,535 +283,14 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
         }
     }
 
-    private ApiEntity create0(UpdateApiEntity api, String userId) throws ApiAlreadyExistsException {
-        return this.create0(api, userId, true, null);
-    }
-
-    private ApiEntity create0(UpdateApiEntity api, String userId, boolean createSystemFolder) throws ApiAlreadyExistsException {
-        return this.create0(api, userId, createSystemFolder, null);
-    }
-
-    private ApiEntity create0(UpdateApiEntity api, String userId, boolean createSystemFolder, JsonNode apiDefinition)
-        throws ApiAlreadyExistsException {
-        try {
-            LOGGER.debug("Create {} for user {}", api, userId);
-
-            String apiId = apiDefinition != null && apiDefinition.has("id") ? apiDefinition.get("id").asText() : null;
-            String id = apiId != null && UUID.fromString(apiId) != null ? apiId : RandomString.generate();
-
-            Optional<Api> checkApi = apiRepository.findById(id);
-            if (checkApi.isPresent()) {
-                throw new ApiAlreadyExistsException(id);
-            }
-
-            // if user changes sharding tags, then check if he is allowed to do it
-            checkShardingTags(api, null);
-
-            // format context-path and check if context path is unique
-            final Collection<VirtualHost> sanitizedVirtualHosts = virtualHostService.sanitizeAndValidate(api.getProxy().getVirtualHosts());
-            api.getProxy().setVirtualHosts(new ArrayList<>(sanitizedVirtualHosts));
-
-            // check endpoints name
-            checkEndpointsName(api);
-
-            // check HC inheritance
-            checkHealthcheckInheritance(api);
-
-            addLoggingMaxDuration(api.getProxy().getLogging());
-
-            // check if there is regex errors in plaintext fields
-            validateRegexfields(api);
-
-            // check policy configurations.
-            checkPolicyConfigurations(api);
-
-            // check primary owner
-            PrimaryOwnerEntity primaryOwner = findPrimaryOwner(apiDefinition, userId);
-
-            if (apiDefinition != null) {
-                apiDefinition = ((ObjectNode) apiDefinition).put("id", id);
-            }
-
-            Api repoApi = convert(id, api, apiDefinition != null ? apiDefinition.toString() : null);
-
-            if (repoApi != null) {
-                repoApi.setId(id);
-                repoApi.setEnvironmentId(GraviteeContext.getCurrentEnvironment());
-                // Set date fields
-                repoApi.setCreatedAt(new Date());
-                repoApi.setUpdatedAt(repoApi.getCreatedAt());
-                // Be sure that lifecycle is set to STOPPED by default and visibility is private
-                repoApi.setLifecycleState(LifecycleState.STOPPED);
-
-                if (api.getVisibility() == null) {
-                    repoApi.setVisibility(Visibility.PRIVATE);
-                } else {
-                    repoApi.setVisibility(Visibility.valueOf(api.getVisibility().toString()));
-                }
-
-                // Add Default groups
-                Set<String> defaultGroups = groupService
-                    .findByEvent(GroupEvent.API_CREATE)
-                    .stream()
-                    .map(GroupEntity::getId)
-                    .collect(toSet());
-                if (!defaultGroups.isEmpty() && repoApi.getGroups() == null) {
-                    repoApi.setGroups(defaultGroups);
-                } else if (!defaultGroups.isEmpty()) {
-                    repoApi.getGroups().addAll(defaultGroups);
-                }
-
-                // if po is a group, add it as a member of the API
-                if (ApiPrimaryOwnerMode.GROUP.name().equals(primaryOwner.getType())) {
-                    if (repoApi.getGroups() == null) {
-                        repoApi.setGroups(new HashSet<>());
-                    }
-                    repoApi.getGroups().add(primaryOwner.getId());
-                }
-
-                repoApi.setApiLifecycleState(ApiLifecycleState.CREATED);
-                if (parameterService.findAsBoolean(Key.API_REVIEW_ENABLED, ParameterReferenceType.ENVIRONMENT)) {
-                    workflowService.create(WorkflowReferenceType.API, id, REVIEW, userId, DRAFT, "");
-                }
-
-                Api createdApi = apiRepository.create(repoApi);
-
-                if (createSystemFolder) {
-                    createSystemFolder(createdApi.getId());
-                }
-
-                // Audit
-                auditService.createApiAuditLog(
-                    createdApi.getId(),
-                    Collections.emptyMap(),
-                    API_CREATED,
-                    createdApi.getCreatedAt(),
-                    null,
-                    createdApi
-                );
-
-                // Add the primary owner of the newly created API
-                if (primaryOwner != null) {
-                    membershipService.addRoleToMemberOnReference(
-                        new MembershipService.MembershipReference(MembershipReferenceType.API, createdApi.getId()),
-                        new MembershipService.MembershipMember(
-                            primaryOwner.getId(),
-                            null,
-                            MembershipMemberType.valueOf(primaryOwner.getType())
-                        ),
-                        new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-                    );
-
-                    // create the default mail notification
-                    final String emailMetadataValue = "${(api.primaryOwner.email)!''}";
-
-                    GenericNotificationConfigEntity notificationConfigEntity = new GenericNotificationConfigEntity();
-                    notificationConfigEntity.setName("Default Mail Notifications");
-                    notificationConfigEntity.setReferenceType(HookScope.API.name());
-                    notificationConfigEntity.setReferenceId(createdApi.getId());
-                    notificationConfigEntity.setHooks(Arrays.stream(ApiHook.values()).map(Enum::name).collect(toList()));
-                    notificationConfigEntity.setNotifier(NotifierServiceImpl.DEFAULT_EMAIL_NOTIFIER_ID);
-                    notificationConfigEntity.setConfig(emailMetadataValue);
-                    genericNotificationConfigService.create(notificationConfigEntity);
-
-                    // create the default mail support metadata
-                    NewApiMetadataEntity newApiMetadataEntity = new NewApiMetadataEntity();
-                    newApiMetadataEntity.setFormat(MetadataFormat.MAIL);
-                    newApiMetadataEntity.setName(DefaultMetadataUpgrader.METADATA_EMAIL_SUPPORT_KEY);
-                    newApiMetadataEntity.setDefaultValue(emailMetadataValue);
-                    newApiMetadataEntity.setValue(emailMetadataValue);
-                    newApiMetadataEntity.setApiId(createdApi.getId());
-                    apiMetadataService.create(newApiMetadataEntity);
-
-                    //TODO add membership log
-                    ApiEntity apiEntity = convert(createdApi, primaryOwner, null);
-                    ApiEntity apiWithMetadata = fetchMetadataForApi(apiEntity);
-
-                    searchEngineService.index(apiWithMetadata, false);
-                    return apiEntity;
-                } else {
-                    LOGGER.error("Unable to create API {} because primary owner role has not been found.", createdApi.getName());
-                    throw new TechnicalManagementException(
-                        "Unable to create API " + createdApi.getName() + " because primary owner role has not been found"
-                    );
-                }
-            } else {
-                LOGGER.error("Unable to create API {} because of previous error.", api.getName());
-                throw new TechnicalManagementException("Unable to create API " + api.getName());
-            }
-        } catch (TechnicalException ex) {
-            LOGGER.error("An error occurs while trying to create {} for user {}", api, userId, ex);
-            throw new TechnicalManagementException("An error occurs while trying create " + api + " for user " + userId, ex);
-        }
-    }
-
-    public PrimaryOwnerEntity findPrimaryOwner(JsonNode apiDefinition, String userId) {
-        ApiPrimaryOwnerMode poMode = ApiPrimaryOwnerMode.valueOf(
-            this.parameterService.find(Key.API_PRIMARY_OWNER_MODE, ParameterReferenceType.ENVIRONMENT)
-        );
-        PrimaryOwnerEntity primaryOwnerFromDefinition = findPrimaryOwnerFromApiDefinition(apiDefinition);
-        switch (poMode) {
-            case USER:
-                if (primaryOwnerFromDefinition == null || ApiPrimaryOwnerMode.GROUP.name().equals(primaryOwnerFromDefinition.getType())) {
-                    return new PrimaryOwnerEntity(userService.findById(userId));
-                }
-                if (ApiPrimaryOwnerMode.USER.name().equals(primaryOwnerFromDefinition.getType())) {
-                    try {
-                        return new PrimaryOwnerEntity(userService.findById(primaryOwnerFromDefinition.getId()));
-                    } catch (UserNotFoundException unfe) {
-                        return new PrimaryOwnerEntity(userService.findById(userId));
-                    }
-                }
-                break;
-            case GROUP:
-                if (primaryOwnerFromDefinition == null) {
-                    return getFirstPoGroupUserBelongsTo(userId);
-                }
-                if (ApiPrimaryOwnerMode.GROUP.name().equals(primaryOwnerFromDefinition.getType())) {
-                    try {
-                        return new PrimaryOwnerEntity(groupService.findById(primaryOwnerFromDefinition.getId()));
-                    } catch (GroupNotFoundException unfe) {
-                        return getFirstPoGroupUserBelongsTo(userId);
-                    }
-                }
-                if (ApiPrimaryOwnerMode.USER.name().equals(primaryOwnerFromDefinition.getType())) {
-                    try {
-                        final String poUserId = primaryOwnerFromDefinition.getId();
-                        userService.findById(poUserId);
-                        final Set<GroupEntity> poGroupsOfPoUser = groupService
-                            .findByUser(poUserId)
-                            .stream()
-                            .filter(group -> group.getApiPrimaryOwner() != null && !group.getApiPrimaryOwner().isEmpty())
-                            .collect(toSet());
-                        if (poGroupsOfPoUser.isEmpty()) {
-                            return getFirstPoGroupUserBelongsTo(userId);
-                        }
-                        return new PrimaryOwnerEntity(poGroupsOfPoUser.iterator().next());
-                    } catch (UserNotFoundException unfe) {
-                        return getFirstPoGroupUserBelongsTo(userId);
-                    }
-                }
-                break;
-            case HYBRID:
-            default:
-                if (primaryOwnerFromDefinition == null) {
-                    return new PrimaryOwnerEntity(userService.findById(userId));
-                }
-                if (ApiPrimaryOwnerMode.GROUP.name().equals(primaryOwnerFromDefinition.getType())) {
-                    try {
-                        return new PrimaryOwnerEntity(groupService.findById(primaryOwnerFromDefinition.getId()));
-                    } catch (GroupNotFoundException unfe) {
-                        try {
-                            return getFirstPoGroupUserBelongsTo(userId);
-                        } catch (NoPrimaryOwnerGroupForUserException ex) {
-                            return new PrimaryOwnerEntity(userService.findById(userId));
-                        }
-                    }
-                }
-                if (ApiPrimaryOwnerMode.USER.name().equals(primaryOwnerFromDefinition.getType())) {
-                    try {
-                        return new PrimaryOwnerEntity(userService.findById(primaryOwnerFromDefinition.getId()));
-                    } catch (UserNotFoundException unfe) {
-                        return new PrimaryOwnerEntity(userService.findById(userId));
-                    }
-                }
-                break;
-        }
-
-        return new PrimaryOwnerEntity(userService.findById(userId));
-    }
-
-    @NotNull
-    private PrimaryOwnerEntity getFirstPoGroupUserBelongsTo(String userId) {
-        final Set<GroupEntity> poGroupsOfCurrentUser = groupService
-            .findByUser(userId)
-            .stream()
-            .filter(group -> !StringUtils.isEmpty(group.getApiPrimaryOwner()))
-            .collect(toSet());
-        if (poGroupsOfCurrentUser.isEmpty()) {
-            throw new NoPrimaryOwnerGroupForUserException(userId);
-        }
-        return new PrimaryOwnerEntity(poGroupsOfCurrentUser.iterator().next());
-    }
-
-    private PrimaryOwnerEntity findPrimaryOwnerFromApiDefinition(JsonNode apiDefinition) {
-        PrimaryOwnerEntity primaryOwnerEntity = null;
-        if (apiDefinition != null && apiDefinition.has("primaryOwner")) {
-            try {
-                primaryOwnerEntity = objectMapper.readValue(apiDefinition.get("primaryOwner").toString(), PrimaryOwnerEntity.class);
-            } catch (JsonProcessingException e) {
-                LOGGER.warn("Cannot parse primary owner from definition, continue with current user", e);
-            }
-        }
-        return primaryOwnerEntity;
-    }
-
-    private void createSystemFolder(String apiId) {
-        NewPageEntity asideSystemFolder = new NewPageEntity();
-        asideSystemFolder.setName(SystemFolderType.ASIDE.folderName());
-        asideSystemFolder.setPublished(true);
-        asideSystemFolder.setType(PageType.SYSTEM_FOLDER);
-        asideSystemFolder.setVisibility(io.gravitee.rest.api.model.Visibility.PUBLIC);
-        pageService.createPage(apiId, asideSystemFolder);
-    }
-
-    private void checkEndpointsName(UpdateApiEntity api) {
-        if (api.getProxy() != null && api.getProxy().getGroups() != null) {
-            for (EndpointGroup group : api.getProxy().getGroups()) {
-                assertEndpointNameNotContainsInvalidCharacters(group.getName());
-                if (group.getEndpoints() != null) {
-                    for (Endpoint endpoint : group.getEndpoints()) {
-                        assertEndpointNameNotContainsInvalidCharacters(endpoint.getName());
-                    }
-                }
-            }
-        }
-    }
-
-    private void checkEndpointsExists(UpdateApiEntity api) {
-        if (api.getProxy().getGroups() == null || api.getProxy().getGroups().isEmpty()) {
-            throw new EndpointMissingException();
-        }
-
-        EndpointGroup endpointGroup = api.getProxy().getGroups().iterator().next();
-        //Is service discovery enabled ?
-        EndpointDiscoveryService endpointDiscoveryService = endpointGroup.getServices() == null
-            ? null
-            : endpointGroup.getServices().get(EndpointDiscoveryService.class);
-        if (
-            (endpointDiscoveryService == null || !endpointDiscoveryService.isEnabled()) &&
-            (endpointGroup.getEndpoints() == null || endpointGroup.getEndpoints().isEmpty())
-        ) {
-            throw new EndpointMissingException();
-        }
-    }
-
-    private void validateHealtcheckSchedule(UpdateApiEntity api) {
-        if (api.getServices() != null) {
-            HealthCheckService healthCheckService = api.getServices().get(HealthCheckService.class);
-            if (healthCheckService != null) {
-                String schedule = healthCheckService.getSchedule();
-                if (schedule != null) {
-                    try {
-                        new CronTrigger(schedule);
-                    } catch (IllegalArgumentException e) {
-                        throw new InvalidDataException(e);
-                    }
-                }
-            }
-        }
-    }
-
-    private void checkHealthcheckInheritance(UpdateApiEntity api) {
-        boolean inherit = false;
-
-        if (api.getProxy() != null && api.getProxy().getGroups() != null) {
-            for (EndpointGroup group : api.getProxy().getGroups()) {
-                if (group.getEndpoints() != null) {
-                    for (Endpoint endpoint : group.getEndpoints()) {
-                        if (endpoint instanceof HttpEndpoint) {
-                            HttpEndpoint httpEndpoint = (HttpEndpoint) endpoint;
-                            if (httpEndpoint.getHealthCheck() != null && httpEndpoint.getHealthCheck().isInherit()) {
-                                inherit = true;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        if (inherit) {
-            //if endpoints are set to inherit HC configuration, this configuration must exists.
-            boolean hcServiceExists = false;
-            if (api.getServices() != null) {
-                HealthCheckService healthCheckService = api.getServices().get(HealthCheckService.class);
-                hcServiceExists = healthCheckService != null;
-            }
-
-            if (!hcServiceExists) {
-                throw new HealthcheckInheritanceException();
-            }
-        }
-    }
-
-    private void assertEndpointNameNotContainsInvalidCharacters(String name) {
-        if (name != null && name.contains(":")) {
-            throw new EndpointNameInvalidException(name);
-        }
-    }
-
-    private void addLoggingMaxDuration(Logging logging) {
-        if (logging != null && !LoggingMode.NONE.equals(logging.getMode())) {
-            Optional<Long> optionalMaxDuration = parameterService
-                .findAll(Key.LOGGING_DEFAULT_MAX_DURATION, Long::valueOf, ParameterReferenceType.ORGANIZATION)
-                .stream()
-                .findFirst();
-            if (optionalMaxDuration.isPresent() && optionalMaxDuration.get() > 0) {
-                long maxEndDate = System.currentTimeMillis() + optionalMaxDuration.get();
-
-                // if no condition set, add one
-                if (logging.getCondition() == null || logging.getCondition().isEmpty()) {
-                    logging.setCondition(String.format(LOGGING_MAX_DURATION_CONDITION, maxEndDate));
-                } else {
-                    Matcher matcher = LOGGING_MAX_DURATION_PATTERN.matcher(logging.getCondition());
-                    if (matcher.matches()) {
-                        String currentDurationAsStr = matcher.group("timestamp");
-                        String before = formatExpression(matcher, "before");
-                        String after = formatExpression(matcher, "after");
-                        try {
-                            final long currentDuration = Long.parseLong(currentDurationAsStr);
-                            if (currentDuration > maxEndDate || (!before.isEmpty() || !after.isEmpty())) {
-                                logging.setCondition(before + String.format(LOGGING_MAX_DURATION_CONDITION, maxEndDate) + after);
-                            }
-                        } catch (NumberFormatException nfe) {
-                            LOGGER.error("Wrong format of the logging condition. Add the default one", nfe);
-                            logging.setCondition(before + String.format(LOGGING_MAX_DURATION_CONDITION, maxEndDate) + after);
-                        }
-                    } else {
-                        logging.setCondition(String.format(LOGGING_MAX_DURATION_CONDITION, maxEndDate) + " && " + logging.getCondition());
-                    }
-                }
-            }
-        }
-    }
-
-    private String formatExpression(final Matcher matcher, final String group) {
-        String matchedExpression = matcher.group(group);
-        final boolean expressionBlank = matchedExpression == null || "".equals(matchedExpression);
-        final boolean after = "after".equals(group);
-
-        String expression;
-        if (after) {
-            if (matchedExpression.startsWith(" && (") && matchedExpression.endsWith(")")) {
-                matchedExpression = matchedExpression.substring(5, matchedExpression.length() - 1);
-            }
-            expression = expressionBlank ? "" : " && (" + matchedExpression + ")";
-            expression = expression.replaceAll("\\(" + LOGGING_DELIMITER_BASE, "\\(");
-        } else {
-            if (matchedExpression.startsWith("(") && matchedExpression.endsWith(") && ")) {
-                matchedExpression = matchedExpression.substring(1, matchedExpression.length() - 5);
-            }
-            expression = expressionBlank ? "" : "(" + matchedExpression + ") && ";
-            expression = expression.replaceAll(LOGGING_DELIMITER_BASE + "\\)", "\\)");
-        }
-        return expression;
-    }
-
     @Override
     public ApiEntity findById(String apiId) {
-        final Api api = this.findApiById(apiId);
-        ApiEntity apiEntity = convert(api, getPrimaryOwner(api), null);
-
-        // Compute entrypoints
-        calculateEntrypoints(apiEntity, api.getEnvironmentId());
-
-        return apiEntity;
+        return super.findById(apiId);
     }
 
     @Override
     public PrimaryOwnerEntity getPrimaryOwner(String apiId) throws TechnicalManagementException {
-        MembershipEntity primaryOwnerMemberEntity = membershipService.getPrimaryOwner(
-            io.gravitee.rest.api.model.MembershipReferenceType.API,
-            apiId
-        );
-        if (primaryOwnerMemberEntity == null) {
-            LOGGER.error("The API {} doesn't have any primary owner.", apiId);
-            throw new TechnicalManagementException("The API " + apiId + " doesn't have any primary owner.");
-        }
-        if (MembershipMemberType.GROUP == primaryOwnerMemberEntity.getMemberType()) {
-            return new PrimaryOwnerEntity(groupService.findById(primaryOwnerMemberEntity.getMemberId()));
-        }
-        return new PrimaryOwnerEntity(userService.findById(primaryOwnerMemberEntity.getMemberId()));
-    }
-
-    private PrimaryOwnerEntity getPrimaryOwner(Api api) throws TechnicalManagementException {
-        return this.getPrimaryOwner(api.getId());
-    }
-
-    private void calculateEntrypoints(ApiEntity api, String environmentId) {
-        List<ApiEntrypointEntity> apiEntrypoints = new ArrayList<>();
-
-        if (api.getProxy() != null) {
-            String defaultEntrypoint = parameterService.find(Key.PORTAL_ENTRYPOINT, environmentId, ParameterReferenceType.ENVIRONMENT);
-            final String scheme = getScheme(defaultEntrypoint);
-            if (api.getTags() != null && !api.getTags().isEmpty()) {
-                List<EntrypointEntity> entrypoints = entrypointService.findByReference(
-                    GraviteeContext.getCurrentOrganization(),
-                    EntrypointReferenceType.ORGANIZATION
-                );
-                entrypoints.forEach(
-                    entrypoint -> {
-                        Set<String> tagEntrypoints = new HashSet<>(Arrays.asList(entrypoint.getTags()));
-                        tagEntrypoints.retainAll(api.getTags());
-
-                        if (tagEntrypoints.size() == entrypoint.getTags().length) {
-                            api
-                                .getProxy()
-                                .getVirtualHosts()
-                                .forEach(
-                                    virtualHost -> {
-                                        String targetHost = (virtualHost.getHost() == null || !virtualHost.isOverrideEntrypoint())
-                                            ? entrypoint.getValue()
-                                            : virtualHost.getHost();
-                                        if (!targetHost.toLowerCase().startsWith("http")) {
-                                            targetHost = scheme + "://" + targetHost;
-                                        }
-                                        apiEntrypoints.add(
-                                            new ApiEntrypointEntity(
-                                                tagEntrypoints,
-                                                DUPLICATE_SLASH_REMOVER
-                                                    .matcher(targetHost + URI_PATH_SEPARATOR + virtualHost.getPath())
-                                                    .replaceAll(URI_PATH_SEPARATOR),
-                                                virtualHost.getHost()
-                                            )
-                                        );
-                                    }
-                                );
-                        }
-                    }
-                );
-            }
-
-            // If empty, get the default entrypoint
-            if (apiEntrypoints.isEmpty()) {
-                api
-                    .getProxy()
-                    .getVirtualHosts()
-                    .forEach(
-                        virtualHost -> {
-                            String targetHost = (virtualHost.getHost() == null || !virtualHost.isOverrideEntrypoint())
-                                ? defaultEntrypoint
-                                : virtualHost.getHost();
-                            if (!targetHost.toLowerCase().startsWith("http")) {
-                                targetHost = scheme + "://" + targetHost;
-                            }
-                            apiEntrypoints.add(
-                                new ApiEntrypointEntity(
-                                    DUPLICATE_SLASH_REMOVER
-                                        .matcher(targetHost + URI_PATH_SEPARATOR + virtualHost.getPath())
-                                        .replaceAll(URI_PATH_SEPARATOR),
-                                    virtualHost.getHost()
-                                )
-                            );
-                        }
-                    );
-            }
-        }
-
-        api.setEntrypoints(apiEntrypoints);
-    }
-
-    private String getScheme(String defaultEntrypoint) {
-        String scheme = "https";
-        if (defaultEntrypoint != null) {
-            try {
-                scheme = new URL(defaultEntrypoint).getProtocol();
-            } catch (MalformedURLException e) {
-                // return default scheme
-            }
-        }
-        return scheme;
+        return super.getPrimaryOwner(apiId);
     }
 
     @Override
@@ -1051,27 +386,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
             final String errorMessage = "An error occurs while trying to search for API ids for user " + userId + ": " + apiQuery;
             LOGGER.error(errorMessage, ex);
             throw new TechnicalManagementException(errorMessage, ex);
-        }
-    }
-
-    private Api findApiById(String apiId) {
-        try {
-            LOGGER.debug("Find API by ID: {}", apiId);
-
-            Optional<Api> optApi = apiRepository.findById(apiId);
-
-            if (GraviteeContext.getCurrentEnvironment() != null) {
-                optApi = optApi.filter(result -> result.getEnvironmentId().equals(GraviteeContext.getCurrentEnvironment()));
-            }
-
-            if (optApi.isPresent()) {
-                return optApi.get();
-            }
-
-            throw new ApiNotFoundException(apiId);
-        } catch (TechnicalException ex) {
-            LOGGER.error("An error occurs while trying to find an API using its ID: {}", apiId, ex);
-            throw new TechnicalManagementException("An error occurs while trying to find an API using its ID: " + apiId, ex);
         }
     }
 
@@ -1345,386 +659,7 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
 
     @Override
     public ApiEntity update(String apiId, UpdateApiEntity updateApiEntity, boolean checkPlans) {
-        try {
-            LOGGER.debug("Update API {}", apiId);
-
-            Optional<Api> optApiToUpdate = apiRepository.findById(apiId);
-            if (!optApiToUpdate.isPresent()) {
-                throw new ApiNotFoundException(apiId);
-            }
-
-            // check if entrypoints are unique
-            final Collection<VirtualHost> sanitizedVirtualHosts = virtualHostService.sanitizeAndValidate(
-                updateApiEntity.getProxy().getVirtualHosts(),
-                apiId
-            );
-            updateApiEntity.getProxy().setVirtualHosts(new ArrayList<>(sanitizedVirtualHosts));
-
-            // check endpoints presence
-            checkEndpointsExists(updateApiEntity);
-
-            // check endpoints name
-            checkEndpointsName(updateApiEntity);
-
-            // check HC inheritance
-            checkHealthcheckInheritance(updateApiEntity);
-
-            // validate HC cron schedule
-            validateHealtcheckSchedule(updateApiEntity);
-
-            // check CORS Allow-origin format
-            checkAllowOriginFormat(updateApiEntity);
-
-            addLoggingMaxDuration(updateApiEntity.getProxy().getLogging());
-
-            // check if there is regex errors in plaintext fields
-            validateRegexfields(updateApiEntity);
-
-            // check policy configurations.
-            checkPolicyConfigurations(updateApiEntity);
-
-            final ApiEntity apiToCheck = convert(optApiToUpdate.get());
-
-            // if user changes sharding tags, then check if he is allowed to do it
-            checkShardingTags(updateApiEntity, apiToCheck);
-
-            // if lifecycle state not provided, set the saved one
-            if (updateApiEntity.getLifecycleState() == null) {
-                updateApiEntity.setLifecycleState(apiToCheck.getLifecycleState());
-            }
-
-            // check lifecycle state
-            checkLifecycleState(updateApiEntity, apiToCheck);
-
-            Set<String> groups = updateApiEntity.getGroups();
-            if (groups != null && !groups.isEmpty()) {
-                // check the existence of groups
-                checkGroupExistence(groups);
-
-                // remove PO group if exists
-                groups = removePOGroups(groups, apiId);
-                updateApiEntity.setGroups(groups);
-            }
-
-            // add a default path, if version is v1
-            if (
-                Objects.equals(updateApiEntity.getGraviteeDefinitionVersion(), DefinitionVersion.V1.getLabel()) &&
-                (updateApiEntity.getPaths() == null || updateApiEntity.getPaths().isEmpty())
-            ) {
-                updateApiEntity.setPaths(singletonMap("/", new ArrayList<>()));
-            }
-
-            if (updateApiEntity.getPlans() == null) {
-                updateApiEntity.setPlans(new ArrayList<>());
-            } else if (checkPlans) {
-                List<Plan> existingPlans = apiToCheck.getPlans();
-                Map<String, String> planStatuses = new HashMap<>();
-                if (existingPlans != null && !existingPlans.isEmpty()) {
-                    planStatuses.putAll(existingPlans.stream().collect(toMap(Plan::getId, Plan::getStatus)));
-                }
-
-                updateApiEntity
-                    .getPlans()
-                    .forEach(
-                        planToUpdate -> {
-                            if (
-                                !planStatuses.containsKey(planToUpdate.getId()) ||
-                                (
-                                    planStatuses.containsKey(planToUpdate.getId()) &&
-                                    planStatuses.get(planToUpdate.getId()).equalsIgnoreCase(PlanStatus.CLOSED.name()) &&
-                                    !planStatuses.get(planToUpdate.getId()).equalsIgnoreCase(planToUpdate.getStatus())
-                                )
-                            ) {
-                                throw new InvalidDataException("Invalid status for plan '" + planToUpdate.getName() + "'");
-                            }
-                        }
-                    );
-            }
-
-            Api apiToUpdate = optApiToUpdate.get();
-
-            if (io.gravitee.rest.api.model.api.ApiLifecycleState.DEPRECATED.equals(updateApiEntity.getLifecycleState())) {
-                planService
-                    .findByApi(apiId)
-                    .forEach(
-                        plan -> {
-                            if (PlanStatus.PUBLISHED.equals(plan.getStatus()) || PlanStatus.STAGING.equals(plan.getStatus())) {
-                                planService.deprecate(plan.getId(), true);
-                                updateApiEntity
-                                    .getPlans()
-                                    .stream()
-                                    .filter(p -> p.getId().equals(plan.getId()))
-                                    .forEach(p -> p.setStatus(PlanStatus.DEPRECATED.name()));
-                            }
-                        }
-                    );
-            }
-
-            Api api = convert(apiId, updateApiEntity, apiToUpdate.getDefinition());
-
-            if (api != null) {
-                api.setId(apiId.trim());
-                api.setUpdatedAt(new Date());
-
-                // Copy fields from existing values
-                api.setEnvironmentId(apiToUpdate.getEnvironmentId());
-                api.setDeployedAt(apiToUpdate.getDeployedAt());
-                api.setCreatedAt(apiToUpdate.getCreatedAt());
-                api.setLifecycleState(apiToUpdate.getLifecycleState());
-                // If no new picture and the current picture url is not the default one, keep the current picture
-                if (
-                    updateApiEntity.getPicture() == null &&
-                    updateApiEntity.getPictureUrl() != null &&
-                    updateApiEntity.getPictureUrl().indexOf("?hash") > 0
-                ) {
-                    api.setPicture(apiToUpdate.getPicture());
-                }
-                if (
-                    updateApiEntity.getBackground() == null &&
-                    updateApiEntity.getBackgroundUrl() != null &&
-                    updateApiEntity.getBackgroundUrl().indexOf("?hash") > 0
-                ) {
-                    api.setBackground(apiToUpdate.getBackground());
-                }
-                if (updateApiEntity.getGroups() == null) {
-                    api.setGroups(apiToUpdate.getGroups());
-                }
-                if (updateApiEntity.getLabels() == null && apiToUpdate.getLabels() != null) {
-                    api.setLabels(new ArrayList<>(new HashSet<>(apiToUpdate.getLabels())));
-                }
-                if (updateApiEntity.getCategories() == null) {
-                    api.setCategories(apiToUpdate.getCategories());
-                }
-
-                if (ApiLifecycleState.DEPRECATED.equals(api.getApiLifecycleState())) {
-                    notifierService.trigger(
-                        ApiHook.API_DEPRECATED,
-                        apiId,
-                        new NotificationParamsBuilder().api(apiToCheck).user(userService.findById(getAuthenticatedUsername())).build()
-                    );
-                }
-
-                Api updatedApi = apiRepository.update(api);
-
-                // Audit
-                auditService.createApiAuditLog(
-                    updatedApi.getId(),
-                    Collections.emptyMap(),
-                    API_UPDATED,
-                    updatedApi.getUpdatedAt(),
-                    apiToUpdate,
-                    updatedApi
-                );
-
-                if (parameterService.findAsBoolean(Key.LOGGING_AUDIT_TRAIL_ENABLED, ParameterReferenceType.ENVIRONMENT)) {
-                    // Audit API logging if option is enabled
-                    auditApiLogging(apiToUpdate, updatedApi);
-                }
-
-                ApiEntity apiEntity = convert(singletonList(updatedApi)).iterator().next();
-                ApiEntity apiWithMetadata = fetchMetadataForApi(apiEntity);
-
-                searchEngineService.index(apiWithMetadata, false);
-
-                return apiEntity;
-            } else {
-                LOGGER.error("Unable to update API {} because of previous error.", apiId);
-                throw new TechnicalManagementException("Unable to update API " + apiId);
-            }
-        } catch (TechnicalException ex) {
-            LOGGER.error("An error occurs while trying to update API {}", apiId, ex);
-            throw new TechnicalManagementException("An error occurs while trying to update API " + apiId, ex);
-        }
-    }
-
-    private String buildApiDefinition(String apiId, String apiDefinition, UpdateApiEntity updateApiEntity) {
-        try {
-            io.gravitee.definition.model.Api updateApiDefinition;
-            if (apiDefinition == null || apiDefinition.isEmpty()) {
-                updateApiDefinition = new io.gravitee.definition.model.Api();
-                updateApiDefinition.setDefinitionVersion(DefinitionVersion.valueOfLabel(updateApiEntity.getGraviteeDefinitionVersion()));
-            } else {
-                updateApiDefinition = objectMapper.readValue(apiDefinition, io.gravitee.definition.model.Api.class);
-            }
-            updateApiDefinition.setId(apiId);
-            updateApiDefinition.setName(updateApiEntity.getName());
-            updateApiDefinition.setVersion(updateApiEntity.getVersion());
-            updateApiDefinition.setProxy(updateApiEntity.getProxy());
-
-            if (StringUtils.isNotEmpty(updateApiEntity.getGraviteeDefinitionVersion())) {
-                updateApiDefinition.setDefinitionVersion(DefinitionVersion.valueOfLabel(updateApiEntity.getGraviteeDefinitionVersion()));
-            }
-
-            if (updateApiEntity.getFlowMode() != null) {
-                updateApiDefinition.setFlowMode(updateApiEntity.getFlowMode());
-            }
-
-            if (updateApiEntity.getPaths() != null) {
-                updateApiDefinition.setPaths(updateApiEntity.getPaths());
-            }
-            if (updateApiEntity.getPathMappings() != null) {
-                updateApiDefinition.setPathMappings(
-                    updateApiEntity
-                        .getPathMappings()
-                        .stream()
-                        .collect(toMap(pathMapping -> pathMapping, pathMapping -> Pattern.compile("")))
-                );
-            }
-            if (updateApiEntity.getFlows() != null) {
-                updateApiDefinition.setFlows(updateApiEntity.getFlows());
-            }
-            if (updateApiEntity.getPlans() != null) {
-                List<Plan> plans = updateApiEntity.getPlans().stream().filter(plan -> plan.getId() != null).collect(toList());
-                updateApiDefinition.setPlans(plans);
-            }
-
-            updateApiDefinition.setServices(updateApiEntity.getServices());
-            updateApiDefinition.setResources(updateApiEntity.getResources());
-            updateApiDefinition.setProperties(updateApiEntity.getProperties());
-            updateApiDefinition.setTags(updateApiEntity.getTags());
-
-            updateApiDefinition.setResponseTemplates(updateApiEntity.getResponseTemplates());
-            return objectMapper.writeValueAsString(updateApiDefinition);
-        } catch (JsonProcessingException jse) {
-            LOGGER.error("Unexpected error while generating API definition", jse);
-            throw new TechnicalManagementException("An error occurs while trying to parse API definition " + jse);
-        }
-    }
-
-    private void checkAllowOriginFormat(UpdateApiEntity updateApiEntity) {
-        if (updateApiEntity.getProxy() != null && updateApiEntity.getProxy().getCors() != null) {
-            final Set<String> accessControlAllowOrigin = updateApiEntity.getProxy().getCors().getAccessControlAllowOrigin();
-            if (accessControlAllowOrigin != null && !accessControlAllowOrigin.isEmpty()) {
-                for (String allowOriginItem : accessControlAllowOrigin) {
-                    if (!CORS_REGEX_PATTERN.matcher(allowOriginItem).matches()) {
-                        if (StringUtils.indexOfAny(allowOriginItem, CORS_REGEX_CHARS) >= 0) {
-                            try {
-                                //the origin could be a regex
-                                Pattern.compile(allowOriginItem);
-                            } catch (PatternSyntaxException e) {
-                                throw new AllowOriginNotAllowedException(allowOriginItem);
-                            }
-                        } else {
-                            throw new AllowOriginNotAllowedException(allowOriginItem);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private void checkShardingTags(final UpdateApiEntity updateApiEntity, final ApiEntity existingAPI) {
-        final Set<String> tagsToUpdate = updateApiEntity.getTags() == null ? new HashSet<>() : updateApiEntity.getTags();
-        final Set<String> updatedTags;
-        if (existingAPI == null) {
-            updatedTags = tagsToUpdate;
-        } else {
-            final Set<String> existingAPITags = existingAPI.getTags() == null ? new HashSet<>() : existingAPI.getTags();
-            updatedTags = existingAPITags.stream().filter(tag -> !tagsToUpdate.contains(tag)).collect(toSet());
-            updatedTags.addAll(tagsToUpdate.stream().filter(tag -> !existingAPITags.contains(tag)).collect(toSet()));
-        }
-        if (updatedTags != null && !updatedTags.isEmpty()) {
-            final Set<String> userTags = tagService.findByUser(
-                getAuthenticatedUsername(),
-                GraviteeContext.getCurrentOrganization(),
-                TagReferenceType.ORGANIZATION
-            );
-            if (!userTags.containsAll(updatedTags)) {
-                final String[] notAllowedTags = updatedTags.stream().filter(tag -> !userTags.contains(tag)).toArray(String[]::new);
-                throw new TagNotAllowedException(notAllowedTags);
-            }
-        }
-    }
-
-    private void checkPolicyConfigurations(final UpdateApiEntity updateApiEntity) {
-        if (updateApiEntity.getPaths() != null) {
-            updateApiEntity
-                .getPaths()
-                .forEach(
-                    (s, rules) ->
-                        rules
-                            .stream()
-                            .filter(Rule::isEnabled)
-                            .map(Rule::getPolicy)
-                            .forEach(policy -> policyService.validatePolicyConfiguration(policy))
-                );
-        }
-        if (updateApiEntity.getFlows() != null) {
-            updateApiEntity
-                .getFlows()
-                .stream()
-                .filter(flow -> flow.getPre() != null)
-                .forEach(
-                    flow -> flow.getPre().stream().filter(Step::isEnabled).forEach(step -> policyService.validatePolicyConfiguration(step))
-                );
-
-            updateApiEntity
-                .getFlows()
-                .stream()
-                .filter(flow -> flow.getPost() != null)
-                .forEach(
-                    flow -> flow.getPost().stream().filter(Step::isEnabled).forEach(step -> policyService.validatePolicyConfiguration(step))
-                );
-        }
-    }
-
-    private void validateRegexfields(final UpdateApiEntity updateApiEntity) {
-        // validate regex on paths
-        if (updateApiEntity.getPaths() != null) {
-            updateApiEntity
-                .getPaths()
-                .forEach(
-                    (path, v) -> {
-                        try {
-                            Pattern.compile(path);
-                        } catch (java.util.regex.PatternSyntaxException pse) {
-                            LOGGER.error("An error occurs while trying to parse the path {}", path, pse);
-                            throw new TechnicalManagementException("An error occurs while trying to parse the path " + path, pse);
-                        }
-                    }
-                );
-        }
-
-        // validate regex on pathMappings
-        if (updateApiEntity.getPathMappings() != null) {
-            updateApiEntity
-                .getPathMappings()
-                .forEach(
-                    pathMapping -> {
-                        try {
-                            Pattern.compile(pathMapping);
-                        } catch (java.util.regex.PatternSyntaxException pse) {
-                            LOGGER.error("An error occurs while trying to parse the path mapping {}", pathMapping, pse);
-                            throw new TechnicalManagementException(
-                                "An error occurs while trying to parse the path mapping" + pathMapping,
-                                pse
-                            );
-                        }
-                    }
-                );
-        }
-    }
-
-    private void checkLifecycleState(final UpdateApiEntity updateApiEntity, final ApiEntity existingAPI) {
-        if (io.gravitee.rest.api.model.api.ApiLifecycleState.DEPRECATED.equals(existingAPI.getLifecycleState())) {
-            throw new LifecycleStateChangeNotAllowedException(updateApiEntity.getLifecycleState().name());
-        }
-        if (existingAPI.getLifecycleState().name().equals(updateApiEntity.getLifecycleState().name())) {
-            return;
-        }
-        if (io.gravitee.rest.api.model.api.ApiLifecycleState.ARCHIVED.equals(existingAPI.getLifecycleState())) {
-            if (!io.gravitee.rest.api.model.api.ApiLifecycleState.ARCHIVED.equals(updateApiEntity.getLifecycleState())) {
-                throw new LifecycleStateChangeNotAllowedException(updateApiEntity.getLifecycleState().name());
-            }
-        } else if (io.gravitee.rest.api.model.api.ApiLifecycleState.UNPUBLISHED.equals(existingAPI.getLifecycleState())) {
-            if (io.gravitee.rest.api.model.api.ApiLifecycleState.CREATED.equals(updateApiEntity.getLifecycleState())) {
-                throw new LifecycleStateChangeNotAllowedException(updateApiEntity.getLifecycleState().name());
-            }
-        } else if (io.gravitee.rest.api.model.api.ApiLifecycleState.CREATED.equals(existingAPI.getLifecycleState())) {
-            if (WorkflowState.IN_REVIEW.equals(existingAPI.getWorkflowState())) {
-                throw new LifecycleStateChangeNotAllowedException(updateApiEntity.getLifecycleState().name());
-            }
-        }
+        return super.update(apiId, updateApiEntity, checkPlans);
     }
 
     @Override
@@ -2035,486 +970,8 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
     }
 
     @Override
-    public String exportAsJson(final String apiId, String exportVersion, String... filteredFields) {
-        ApiEntity apiEntity = findById(apiId);
-        // set metadata for serialize process
-        Map<String, Object> metadata = new HashMap<>();
-        metadata.put(ApiSerializer.METADATA_EXPORT_VERSION, exportVersion);
-        metadata.put(ApiSerializer.METADATA_FILTERED_FIELDS_LIST, Arrays.asList(filteredFields));
-        apiEntity.setMetadata(metadata);
-
-        try {
-            return objectMapper.writeValueAsString(apiEntity);
-        } catch (final Exception e) {
-            LOGGER.error("An error occurs while trying to JSON serialize the API {}", apiEntity, e);
-        }
-        return "";
-    }
-
-    @Override
-    public ApiEntity createWithImportedDefinition(ApiEntity apiEntity, String apiDefinitionOrURL, String userId) {
-        String apiDefinition = fetchApiDefinitionContentFromURL(apiDefinitionOrURL);
-        try {
-            // Read the whole definition
-            final JsonNode jsonNode = objectMapper.readTree(apiDefinition);
-            UpdateApiEntity importedApi = this.convertToEntity(apiDefinition, jsonNode);
-            ApiEntity createdApiEntity = create0(importedApi, userId, false, jsonNode);
-            createPageAndMedia(createdApiEntity, jsonNode);
-            updateApiReferences(createdApiEntity, jsonNode);
-            return createdApiEntity;
-        } catch (JsonProcessingException e) {
-            LOGGER.error("An error occurs while trying to JSON deserialize the API {}", apiDefinition, e);
-            throw new TechnicalManagementException("An error occurs while trying to JSON deserialize the API definition.");
-        }
-    }
-
-    private void createPageAndMedia(ApiEntity createdApiEntity, JsonNode jsonNode) {
-        final JsonNode apiMedia = jsonNode.path("apiMedia");
-        if (apiMedia != null && apiMedia.isArray()) {
-            for (JsonNode media : apiMedia) {
-                mediaService.createWithDefinition(createdApiEntity.getId(), media.toString());
-            }
-        }
-
-        final JsonNode pages = jsonNode.path("pages");
-        if (pages != null && pages.isArray()) {
-            for (JsonNode page : pages) {
-                PageEntity pageEntity = pageService.createWithDefinition(createdApiEntity.getId(), page.toString());
-                ((ObjectNode) page).put("id", pageEntity.getId());
-            }
-        }
-
-        List<PageEntity> search = pageService.search(
-            new PageQuery.Builder()
-                .api(createdApiEntity.getId())
-                .name(SystemFolderType.ASIDE.folderName())
-                .type(PageType.SYSTEM_FOLDER)
-                .build()
-        );
-        if (search.isEmpty()) {
-            createSystemFolder(createdApiEntity.getId());
-        }
-    }
-
-    @Override
-    public ApiEntity updateWithImportedDefinition(ApiEntity apiEntity, String apiDefinitionOrURL, String userId) {
-        String apiDefinition = fetchApiDefinitionContentFromURL(apiDefinitionOrURL);
-        try {
-            // Read the whole definition
-            final JsonNode jsonNode = objectMapper.readTree(apiDefinition);
-            UpdateApiEntity importedApi = this.convertToEntity(apiDefinition, jsonNode);
-            ApiEntity updatedApiEntity = update(apiEntity.getId(), importedApi);
-            updateApiReferences(updatedApiEntity, jsonNode);
-            return updatedApiEntity;
-        } catch (JsonProcessingException e) {
-            LOGGER.error("An error occurs while trying to JSON deserialize the API {}", apiDefinition, e);
-            throw new TechnicalManagementException("An error occurs while trying to JSON deserialize the API definition.");
-        }
-    }
-
-    private UpdateApiEntity convertToEntity(String apiDefinition, JsonNode jsonNode) throws JsonProcessingException {
-        final UpdateApiEntity importedApi = objectMapper
-            // because definition could contains other values than the api itself (pages, members)
-            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-            .readValue(apiDefinition, UpdateApiEntity.class);
-
-        // Initialize with a default path
-        if (
-            Objects.equals(importedApi.getGraviteeDefinitionVersion(), DefinitionVersion.V1.getLabel()) &&
-            (importedApi.getPaths() == null || importedApi.getPaths().isEmpty())
-        ) {
-            importedApi.setPaths(Collections.singletonMap("/", new ArrayList<>()));
-        }
-
-        //create group if not exist & replace groupName by groupId
-        if (importedApi.getGroups() != null) {
-            Set<String> groupNames = new HashSet<>(importedApi.getGroups());
-            importedApi.getGroups().clear();
-            for (String name : groupNames) {
-                List<GroupEntity> groupEntities = groupService.findByName(name);
-                GroupEntity group;
-                if (groupEntities.isEmpty()) {
-                    NewGroupEntity newGroupEntity = new NewGroupEntity();
-                    newGroupEntity.setName(name);
-                    group = groupService.create(newGroupEntity);
-                } else {
-                    group = groupEntities.get(0);
-                }
-                importedApi.getGroups().add(group.getId());
-            }
-        }
-
-        // Views & Categories
-        // Before 3.0.2, API 'categories' were called 'views'. This is for compatibility.
-        final JsonNode viewsDefinition = jsonNode.path("views");
-        if (viewsDefinition != null && viewsDefinition.isArray()) {
-            Set<String> categories = new HashSet<>();
-            for (JsonNode viewNode : viewsDefinition) {
-                categories.add(viewNode.asText());
-            }
-            importedApi.setCategories(categories);
-        }
-
-        return importedApi;
-    }
-
-    private void updateApiReferences(ApiEntity createdOrUpdatedApiEntity, JsonNode jsonNode) throws JsonProcessingException {
-        // Members
-        final JsonNode membersToImport = jsonNode.path("members");
-        if (membersToImport != null && membersToImport.isArray()) {
-            // get current members of the api
-            Set<MemberToImport> membersAlreadyPresent = membershipService
-                .getMembersByReference(MembershipReferenceType.API, createdOrUpdatedApiEntity.getId())
-                .stream()
-                .map(
-                    member -> {
-                        UserEntity userEntity = userService.findById(member.getId());
-                        return new MemberToImport(
-                            userEntity.getSource(),
-                            userEntity.getSourceId(),
-                            member.getRoles().stream().map(RoleEntity::getId).collect(Collectors.toList()),
-                            null
-                        );
-                    }
-                )
-                .collect(toSet());
-            // get the current PO
-            RoleEntity poRole = roleService.findPrimaryOwnerRoleByOrganization(GraviteeContext.getCurrentOrganization(), RoleScope.API);
-            if (poRole != null) {
-                String poRoleId = poRole.getId();
-                MemberToImport currentPo = membersAlreadyPresent
-                    .stream()
-                    .filter(memberToImport -> memberToImport.getRoles().contains(poRoleId))
-                    .findFirst()
-                    .orElse(new MemberToImport());
-
-                List<String> roleUsedInTransfert = null;
-                MemberToImport futurePO = null;
-
-                // upsert members
-                for (final JsonNode memberNode : membersToImport) {
-                    MemberToImport memberToImport = objectMapper.readValue(memberNode.toString(), MemberToImport.class);
-                    String roleToAdd = memberToImport.getRole();
-                    List<String> rolesToImport = memberToImport.getRoles();
-                    if (roleToAdd != null && !roleToAdd.isEmpty()) {
-                        if (rolesToImport == null) {
-                            rolesToImport = new ArrayList<>();
-                            memberToImport.setRoles(rolesToImport);
-                        }
-                        Optional<RoleEntity> optRoleToAddEntity = roleService.findByScopeAndName(RoleScope.API, roleToAdd);
-                        if (optRoleToAddEntity.isPresent()) {
-                            rolesToImport.add(optRoleToAddEntity.get().getId());
-                        } else {
-                            LOGGER.warn("Role {} does not exist", roleToAdd);
-                        }
-                    }
-                    if (rolesToImport != null) {
-                        rolesToImport.sort(Comparator.naturalOrder());
-                    }
-                    boolean presentWithSameRole =
-                        memberToImport.getRoles() != null &&
-                        !memberToImport.getRoles().isEmpty() &&
-                        membersAlreadyPresent
-                            .stream()
-                            .anyMatch(
-                                m -> {
-                                    m.getRoles().sort(Comparator.naturalOrder());
-                                    return (
-                                        m.getRoles().equals(memberToImport.getRoles()) &&
-                                        (
-                                            m.getSourceId().equals(memberToImport.getSourceId()) &&
-                                            m.getSource().equals(memberToImport.getSource())
-                                        )
-                                    );
-                                }
-                            );
-
-                    // add/update members if :
-                    //  - not already present with the same role
-                    //  - not the new PO
-                    //  - not the current PO
-                    if (
-                        !presentWithSameRole &&
-                        (
-                            memberToImport.getRoles() != null &&
-                            !memberToImport.getRoles().isEmpty() &&
-                            !memberToImport.getRoles().contains(poRoleId)
-                        ) &&
-                        !(
-                            memberToImport.getSourceId().equals(currentPo.getSourceId()) &&
-                            memberToImport.getSource().equals(currentPo.getSource())
-                        )
-                    ) {
-                        try {
-                            UserEntity userEntity = userService.findBySource(
-                                memberToImport.getSource(),
-                                memberToImport.getSourceId(),
-                                false
-                            );
-
-                            rolesToImport.forEach(
-                                role ->
-                                    membershipService.addRoleToMemberOnReference(
-                                        MembershipReferenceType.API,
-                                        createdOrUpdatedApiEntity.getId(),
-                                        MembershipMemberType.USER,
-                                        userEntity.getId(),
-                                        role
-                                    )
-                            );
-                        } catch (UserNotFoundException unfe) {}
-                    }
-
-                    // get the future role of the current PO
-                    if (
-                        currentPo.getSourceId().equals(memberToImport.getSourceId()) &&
-                        currentPo.getSource().equals(memberToImport.getSource()) &&
-                        !rolesToImport.contains(poRoleId)
-                    ) {
-                        roleUsedInTransfert = rolesToImport;
-                    }
-
-                    if (rolesToImport.contains(poRoleId)) {
-                        futurePO = memberToImport;
-                    }
-                }
-
-                // transfer the ownership
-                if (
-                    futurePO != null &&
-                    !(currentPo.getSource().equals(futurePO.getSource()) && currentPo.getSourceId().equals(futurePO.getSourceId()))
-                ) {
-                    try {
-                        UserEntity userEntity = userService.findBySource(futurePO.getSource(), futurePO.getSourceId(), false);
-                        List<RoleEntity> roleEntity = null;
-                        if (roleUsedInTransfert != null && !roleUsedInTransfert.isEmpty()) {
-                            roleEntity = roleUsedInTransfert.stream().map(roleService::findById).collect(Collectors.toList());
-                        }
-                        membershipService.transferApiOwnership(
-                            createdOrUpdatedApiEntity.getId(),
-                            new MembershipService.MembershipMember(userEntity.getId(), null, MembershipMemberType.USER),
-                            roleEntity
-                        );
-                    } catch (UserNotFoundException unfe) {}
-                }
-            }
-        }
-
-        //Pages
-        final JsonNode pagesDefinition = jsonNode.path("pages");
-        if (pagesDefinition != null && pagesDefinition.isArray()) {
-            List<PageEntity> pagesList = objectMapper.readValue(
-                pagesDefinition.toString(),
-                objectMapper.getTypeFactory().constructCollectionType(List.class, PageEntity.class)
-            );
-            PageEntityTreeNode documentationTree = new PageEntityTreeNode(new PageEntity());
-            documentationTree.appendListToTree(pagesList);
-            createOrUpdateChildrenPages(createdOrUpdatedApiEntity.getId(), null, documentationTree.children);
-        }
-
-        //Plans
-        final JsonNode plansDefinition = jsonNode.path("plans");
-        if (plansDefinition != null && plansDefinition.isArray()) {
-            for (JsonNode planNode : plansDefinition) {
-                if (planNode.has("id")) {
-                    try {
-                        PlanEntity plan = planService.findById(planNode.get("id").asText());
-                        UpdatePlanEntity updatePlanEntity = objectMapper.readValue(planNode.toString(), UpdatePlanEntity.class);
-                        updatePlanEntity.setId(plan.getId());
-                        planService.update(updatePlanEntity);
-                    } catch (PlanNotFoundException npe) {
-                        NewPlanEntity newPlanEntity = objectMapper.readValue(planNode.toString(), NewPlanEntity.class);
-                        newPlanEntity.setApi(createdOrUpdatedApiEntity.getId());
-                        planService.create(newPlanEntity);
-                    }
-                } else {
-                    PlanQuery query = new PlanQuery.Builder()
-                        .api(createdOrUpdatedApiEntity.getId())
-                        .name(planNode.get("name").asText())
-                        .security(PlanSecurityType.valueOf(planNode.get("security").asText().toUpperCase()))
-                        .build();
-                    List<PlanEntity> planEntities = planService
-                        .search(query)
-                        .stream()
-                        .filter(planEntity -> !PlanStatus.CLOSED.equals(planEntity.getStatus()))
-                        .collect(toList());
-                    if (planEntities.isEmpty()) {
-                        NewPlanEntity newPlanEntity = objectMapper.readValue(planNode.toString(), NewPlanEntity.class);
-                        newPlanEntity.setApi(createdOrUpdatedApiEntity.getId());
-                        planService.create(newPlanEntity);
-                    } else if (planEntities.size() == 1) {
-                        UpdatePlanEntity updatePlanEntity = objectMapper.readValue(planNode.toString(), UpdatePlanEntity.class);
-                        updatePlanEntity.setId(planEntities.iterator().next().getId());
-                        planService.update(updatePlanEntity);
-                    } else {
-                        LOGGER.error(
-                            "Not able to identify the plan to update: {}. Too much plan with the same name",
-                            planNode.get("name").asText()
-                        );
-                        throw new TechnicalManagementException(
-                            "Not able to identify the plan to update: " +
-                            planNode.get("name").asText() +
-                            ". Too much plan with the same name"
-                        );
-                    }
-                }
-            }
-        }
-        // Metadata
-        final JsonNode metadataDefinition = jsonNode.path("metadata");
-        if (metadataDefinition != null && metadataDefinition.isArray()) {
-            try {
-                for (JsonNode metadataNode : metadataDefinition) {
-                    UpdateApiMetadataEntity updateApiMetadataEntity = objectMapper.readValue(
-                        metadataNode.toString(),
-                        UpdateApiMetadataEntity.class
-                    );
-                    updateApiMetadataEntity.setApiId(createdOrUpdatedApiEntity.getId());
-                    apiMetadataService.update(updateApiMetadataEntity);
-                }
-            } catch (Exception ex) {
-                LOGGER.error("An error occurs while creating API metadata", ex);
-                throw new TechnicalManagementException("An error occurs while creating API Metadata", ex);
-            }
-        }
-    }
-
-    private String fetchApiDefinitionContentFromURL(String apiDefinitionOrURL) {
-        if (apiDefinitionOrURL.toUpperCase().startsWith("HTTP")) {
-            UrlSanitizerUtils.checkAllowed(
-                apiDefinitionOrURL,
-                importConfiguration.getImportWhitelist(),
-                importConfiguration.isAllowImportFromPrivate()
-            );
-            Buffer buffer = httpClientService.request(HttpMethod.GET, apiDefinitionOrURL, null, null, null);
-            return buffer.toString();
-        }
-        return apiDefinitionOrURL;
-    }
-
-    class PageEntityTreeNode {
-
-        PageEntity data;
-        PageEntityTreeNode parent;
-        List<PageEntityTreeNode> children;
-
-        public PageEntityTreeNode(PageEntity data) {
-            this.data = data;
-            this.children = new LinkedList<>();
-        }
-
-        public PageEntityTreeNode addChild(PageEntity child) {
-            PageEntityTreeNode childNode = new PageEntityTreeNode(child);
-            childNode.parent = this;
-            this.children.add(childNode);
-            return childNode;
-        }
-
-        private PageEntityTreeNode findById(String id) {
-            if (id.equals(data.getId())) {
-                return this;
-            }
-            for (PageEntityTreeNode child : children) {
-                PageEntityTreeNode result = child.findById(id);
-                if (result != null) {
-                    return result;
-                }
-            }
-            return null;
-        }
-
-        public void appendListToTree(List<PageEntity> pagesList) {
-            List<PageEntity> orphans = new ArrayList<>();
-            for (PageEntity newPage : pagesList) {
-                if (newPage.getParentId() == null || newPage.getParentId().isEmpty()) {
-                    this.addChild(newPage);
-                } else {
-                    PageEntityTreeNode parentNode = this.findById(newPage.getParentId());
-                    if (parentNode != null) {
-                        parentNode.addChild(newPage);
-                    } else {
-                        orphans.add(newPage);
-                    }
-                }
-            }
-            if (!orphans.isEmpty() && orphans.size() < pagesList.size()) {
-                appendListToTree(orphans);
-            }
-        }
-    }
-
-    private void createOrUpdateChildrenPages(String apiId, String parentId, List<PageEntityTreeNode> children) {
-        for (final PageEntityTreeNode child : children) {
-            PageEntity pageEntityToImport = child.data;
-            pageEntityToImport.setParentId(parentId);
-
-            PageEntity createdOrUpdatedPage = null;
-            if (pageEntityToImport.getId() != null) {
-                createdOrUpdatedPage = pageService.findById(pageEntityToImport.getId());
-            } else {
-                PageQuery query = new PageQuery.Builder()
-                    .api(apiId)
-                    .name(pageEntityToImport.getName())
-                    .type(PageType.valueOf(pageEntityToImport.getType()))
-                    .build();
-
-                List<PageEntity> pages = pageService.search(query);
-                if (pages.size() == 1) {
-                    createdOrUpdatedPage = pages.get(0);
-                } else if (pages.size() > 1) {
-                    LOGGER.error(
-                        "Not able to identify the page to update: {}. Too much pages with the same name",
-                        pageEntityToImport.getName()
-                    );
-                    throw new TechnicalManagementException(
-                        "Not able to identify the page to update: " + pageEntityToImport.getName() + ". Too much pages with the same name"
-                    );
-                }
-            }
-
-            if (createdOrUpdatedPage == null) {
-                NewPageEntity newPage = new NewPageEntity();
-                newPage.setConfiguration(pageEntityToImport.getConfiguration());
-                newPage.setContent(pageEntityToImport.getContent());
-                newPage.setExcludedAccessControls(pageEntityToImport.isExcludedAccessControls());
-                newPage.setAccessControls(pageEntityToImport.getAccessControls());
-                newPage.setHomepage(pageEntityToImport.isHomepage());
-                newPage.setLastContributor(pageEntityToImport.getLastContributor());
-                newPage.setName(pageEntityToImport.getName());
-                newPage.setOrder(pageEntityToImport.getOrder());
-                newPage.setParentId(pageEntityToImport.getParentId());
-                newPage.setPublished(pageEntityToImport.isPublished());
-                newPage.setSource(pageEntityToImport.getSource());
-                newPage.setType(PageType.valueOf(pageEntityToImport.getType()));
-                newPage.setAttachedMedia(pageEntityToImport.getAttachedMedia());
-                createdOrUpdatedPage = pageService.createPage(apiId, newPage);
-            } else {
-                UpdatePageEntity updatePageEntity = new UpdatePageEntity();
-                updatePageEntity.setConfiguration(pageEntityToImport.getConfiguration());
-                updatePageEntity.setContent(pageEntityToImport.getContent());
-                updatePageEntity.setExcludedAccessControls(pageEntityToImport.isExcludedAccessControls());
-                updatePageEntity.setAccessControls(pageEntityToImport.getAccessControls());
-                updatePageEntity.setHomepage(pageEntityToImport.isHomepage());
-                updatePageEntity.setLastContributor(pageEntityToImport.getLastContributor());
-                updatePageEntity.setName(pageEntityToImport.getName());
-                updatePageEntity.setOrder(pageEntityToImport.getOrder());
-                updatePageEntity.setParentId(pageEntityToImport.getParentId());
-                updatePageEntity.setPublished(pageEntityToImport.isPublished());
-                updatePageEntity.setSource(pageEntityToImport.getSource());
-                updatePageEntity.setAttachedMedia(pageEntityToImport.getAttachedMedia());
-
-                createdOrUpdatedPage = pageService.update(createdOrUpdatedPage.getId(), updatePageEntity);
-            }
-
-            if (child.children != null && !child.children.isEmpty()) {
-                this.createOrUpdateChildrenPages(apiId, createdOrUpdatedPage.getId(), child.children);
-            }
-        }
-    }
-
-    @Override
     public InlinePictureEntity getPicture(String apiId) {
-        Api api = this.findApiById(apiId);
+        Api api = this.findApiById(apiId, GraviteeContext.getCurrentEnvironment());
         InlinePictureEntity imageEntity = new InlinePictureEntity();
         if (api.getPicture() != null) {
             String[] parts = api.getPicture().split(";", 2);
@@ -2548,7 +1005,7 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
 
     @Override
     public InlinePictureEntity getBackground(String apiId) {
-        Api api = this.findApiById(apiId);
+        Api api = this.findApiById(apiId, GraviteeContext.getCurrentEnvironment());
         InlinePictureEntity imageEntity = new InlinePictureEntity();
         if (api.getBackground() != null) {
             String[] parts = api.getBackground().split(";", 2);
@@ -2827,66 +1284,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
     }
 
     @Override
-    public ApiEntity duplicate(final ApiEntity apiEntity, final DuplicateApiEntity duplicateApiEntity) {
-        requireNonNull(apiEntity, "Missing ApiEntity");
-        final String apiId = apiEntity.getId();
-        LOGGER.debug("Duplicate API {}", apiId);
-
-        final UpdateApiEntity newApiEntity = convert(apiEntity);
-        final Proxy proxy = apiEntity.getProxy();
-        proxy.setVirtualHosts(singletonList(new VirtualHost(duplicateApiEntity.getContextPath())));
-        newApiEntity.setProxy(proxy);
-        newApiEntity.setVersion(duplicateApiEntity.getVersion() == null ? apiEntity.getVersion() : duplicateApiEntity.getVersion());
-
-        if (duplicateApiEntity.getFilteredFields().contains("groups")) {
-            newApiEntity.setGroups(null);
-        } else {
-            newApiEntity.setGroups(apiEntity.getGroups());
-        }
-        final ApiEntity duplicatedApi = create0(newApiEntity, getAuthenticatedUsername(), false);
-
-        if (!duplicateApiEntity.getFilteredFields().contains("members")) {
-            final Set<MembershipEntity> membershipsToDuplicate = membershipService.getMembershipsByReference(
-                io.gravitee.rest.api.model.MembershipReferenceType.API,
-                apiId
-            );
-            RoleEntity primaryOwnerRole = roleService.findPrimaryOwnerRoleByOrganization(
-                GraviteeContext.getCurrentOrganization(),
-                RoleScope.API
-            );
-            if (primaryOwnerRole != null) {
-                String primaryOwnerRoleId = primaryOwnerRole.getId();
-                membershipsToDuplicate.forEach(
-                    membership -> {
-                        String roleId = membership.getRoleId();
-                        if (!primaryOwnerRoleId.equals(roleId)) {
-                            membershipService.addRoleToMemberOnReference(
-                                io.gravitee.rest.api.model.MembershipReferenceType.API,
-                                duplicatedApi.getId(),
-                                membership.getMemberType(),
-                                membership.getMemberId(),
-                                roleId
-                            );
-                        }
-                    }
-                );
-            }
-        }
-
-        if (!duplicateApiEntity.getFilteredFields().contains("pages")) {
-            final List<PageEntity> pages = pageService.search(new PageQuery.Builder().api(apiId).build(), true);
-            pages.forEach(page -> pageService.create(duplicatedApi.getId(), page));
-        }
-
-        if (!duplicateApiEntity.getFilteredFields().contains("plans")) {
-            final Set<PlanEntity> plans = planService.findByApi(apiId);
-            plans.forEach(plan -> planService.create(duplicatedApi.getId(), plan));
-        }
-
-        return duplicatedApi;
-    }
-
-    @Override
     public boolean hasHealthCheckEnabled(ApiEntity api, boolean mustBeEnabledOnAllEndpoints) {
         boolean globalHC =
             api.getServices() != null &&
@@ -2927,26 +1324,7 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
 
     @Override
     public ApiEntity fetchMetadataForApi(ApiEntity apiEntity) {
-        List<ApiMetadataEntity> metadataList = apiMetadataService.findAllByApi(apiEntity.getId());
-        final Map<String, Object> mapMetadata = new HashMap<>(metadataList.size());
-
-        metadataList.forEach(
-            metadata -> mapMetadata.put(metadata.getKey(), metadata.getValue() == null ? metadata.getDefaultValue() : metadata.getValue())
-        );
-
-        String decodedValue =
-            this.notificationTemplateService.resolveInlineTemplateWithParam(
-                    apiEntity.getId(),
-                    new StringReader(mapMetadata.toString()),
-                    Collections.singletonMap("api", apiEntity)
-                );
-        Map<String, Object> metadataDecoded = Arrays
-            .stream(decodedValue.substring(1, decodedValue.length() - 1).split(", "))
-            .map(entry -> entry.split("="))
-            .collect(Collectors.toMap(entry -> entry[0], entry -> entry.length > 1 ? entry[1] : ""));
-        apiEntity.setMetadata(metadataDecoded);
-
-        return apiEntity;
+        return super.fetchMetadataForApi(apiEntity);
     }
 
     @Override
@@ -2995,33 +1373,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
             LOGGER.error("An error occurs while trying to remove group {} from API {}: {}", group, apiId, ex);
             throw new TechnicalManagementException("An error occurs while trying to remove group " + group + " from API " + apiId, ex);
         }
-    }
-
-    private UpdateApiEntity convert(final ApiEntity apiEntity) {
-        final UpdateApiEntity updateApiEntity = new UpdateApiEntity();
-        updateApiEntity.setDescription(apiEntity.getDescription());
-        updateApiEntity.setName(apiEntity.getName());
-        updateApiEntity.setVersion(apiEntity.getVersion());
-        updateApiEntity.setGraviteeDefinitionVersion(apiEntity.getGraviteeDefinitionVersion());
-        updateApiEntity.setGroups(apiEntity.getGroups());
-        updateApiEntity.setLabels(apiEntity.getLabels());
-        updateApiEntity.setLifecycleState(apiEntity.getLifecycleState());
-        updateApiEntity.setPicture(apiEntity.getPicture());
-        updateApiEntity.setBackground(apiEntity.getBackground());
-        updateApiEntity.setProperties(apiEntity.getProperties());
-        updateApiEntity.setProxy(apiEntity.getProxy());
-        updateApiEntity.setResources(apiEntity.getResources());
-        updateApiEntity.setResponseTemplates(apiEntity.getResponseTemplates());
-        updateApiEntity.setServices(apiEntity.getServices());
-        updateApiEntity.setTags(apiEntity.getTags());
-        updateApiEntity.setCategories(apiEntity.getCategories());
-        updateApiEntity.setVisibility(apiEntity.getVisibility());
-        updateApiEntity.setPaths(apiEntity.getPaths());
-        updateApiEntity.setFlows(apiEntity.getFlows());
-        updateApiEntity.setPathMappings(apiEntity.getPathMappings());
-        updateApiEntity.setDisableMembershipNotifications(apiEntity.isDisableMembershipNotifications());
-        updateApiEntity.setPlans(apiEntity.getPlans());
-        return updateApiEntity;
     }
 
     private ApiEntity updateWorkflowReview(
@@ -3165,7 +1516,7 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
 
     private void removeTag(String apiId, String tagId) throws TechnicalManagementException {
         try {
-            Api api = this.findApiById(apiId);
+            Api api = this.findApiById(apiId, GraviteeContext.getCurrentEnvironment());
             Api previousApi = new Api(api);
             final io.gravitee.definition.model.Api apiDefinition = objectMapper.readValue(
                 api.getDefinition(),
@@ -3190,7 +1541,7 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
             Api previousApi = new Api(api);
             api.setUpdatedAt(new Date());
             api.setLifecycleState(lifecycleState);
-            ApiEntity apiEntity = convert(apiRepository.update(api), getPrimaryOwner(api), null);
+            ApiEntity apiEntity = convert(apiRepository.update(api), getPrimaryOwner(api.getId()), null);
             // Audit
             auditService.createApiAuditLog(apiId, Collections.emptyMap(), API_UPDATED, api.getUpdatedAt(), previousApi, api);
 
@@ -3212,69 +1563,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
             return apiEntity;
         } else {
             throw new ApiNotFoundException(apiId);
-        }
-    }
-
-    private void auditApiLogging(Api apiToUpdate, Api apiUpdated) {
-        try {
-            // get old logging configuration
-            io.gravitee.definition.model.Api apiToUpdateDefinition = objectMapper.readValue(
-                apiToUpdate.getDefinition(),
-                io.gravitee.definition.model.Api.class
-            );
-            Logging loggingToUpdate = apiToUpdateDefinition.getProxy().getLogging();
-
-            // get new logging configuration
-            io.gravitee.definition.model.Api apiUpdatedDefinition = objectMapper.readValue(
-                apiUpdated.getDefinition(),
-                io.gravitee.definition.model.Api.class
-            );
-            Logging loggingUpdated = apiUpdatedDefinition.getProxy().getLogging();
-
-            // no changes for logging configuration, continue
-            if (
-                loggingToUpdate == loggingUpdated ||
-                (
-                    loggingToUpdate != null &&
-                    loggingUpdated != null &&
-                    Objects.equals(loggingToUpdate.getMode(), loggingUpdated.getMode()) &&
-                    Objects.equals(loggingToUpdate.getCondition(), loggingUpdated.getCondition())
-                )
-            ) {
-                return;
-            }
-
-            // determine the audit event type
-            Api.AuditEvent auditEvent;
-            if (
-                (loggingToUpdate == null || loggingToUpdate.getMode().equals(LoggingMode.NONE)) &&
-                (!loggingUpdated.getMode().equals(LoggingMode.NONE))
-            ) {
-                auditEvent = Api.AuditEvent.API_LOGGING_ENABLED;
-            } else if (
-                (loggingToUpdate != null && !loggingToUpdate.getMode().equals(LoggingMode.NONE)) &&
-                (loggingUpdated.getMode().equals(LoggingMode.NONE))
-            ) {
-                auditEvent = Api.AuditEvent.API_LOGGING_DISABLED;
-            } else {
-                auditEvent = Api.AuditEvent.API_LOGGING_UPDATED;
-            }
-
-            // Audit
-            auditService.createApiAuditLog(
-                apiUpdated.getId(),
-                Collections.emptyMap(),
-                auditEvent,
-                new Date(),
-                loggingToUpdate,
-                loggingUpdated
-            );
-        } catch (Exception ex) {
-            LOGGER.error("An error occurs while auditing API logging configuration for API: {}", apiUpdated.getId(), ex);
-            throw new TechnicalManagementException(
-                "An error occurs while auditing API logging configuration for API: " + apiUpdated.getId(),
-                ex
-            );
         }
     }
 
@@ -3447,48 +1735,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
         return apiEntity;
     }
 
-    private Api convert(String apiId, UpdateApiEntity updateApiEntity, String apiDefinition) {
-        Api api = new Api();
-        api.setId(apiId);
-        if (updateApiEntity.getVisibility() != null) {
-            api.setVisibility(Visibility.valueOf(updateApiEntity.getVisibility().toString()));
-        }
-
-        api.setVersion(updateApiEntity.getVersion().trim());
-        api.setName(updateApiEntity.getName().trim());
-        api.setDescription(updateApiEntity.getDescription().trim());
-        api.setPicture(updateApiEntity.getPicture());
-        api.setBackground(updateApiEntity.getBackground());
-
-        api.setDefinition(buildApiDefinition(apiId, apiDefinition, updateApiEntity));
-
-        final Set<String> apiCategories = updateApiEntity.getCategories();
-        if (apiCategories != null) {
-            final List<CategoryEntity> categories = categoryService.findAll();
-            final Set<String> newApiCategories = new HashSet<>(apiCategories.size());
-            for (final String apiCategory : apiCategories) {
-                final Optional<CategoryEntity> optionalCategory = categories
-                    .stream()
-                    .filter(c -> apiCategory.equals(c.getKey()) || apiCategory.equals(c.getId()))
-                    .findAny();
-                optionalCategory.ifPresent(category -> newApiCategories.add(category.getId()));
-            }
-            api.setCategories(newApiCategories);
-        }
-
-        if (updateApiEntity.getLabels() != null) {
-            api.setLabels(new ArrayList<>(new HashSet<>(updateApiEntity.getLabels())));
-        }
-
-        api.setGroups(updateApiEntity.getGroups());
-        api.setDisableMembershipNotifications(updateApiEntity.isDisableMembershipNotifications());
-
-        if (updateApiEntity.getLifecycleState() != null) {
-            api.setApiLifecycleState(ApiLifecycleState.valueOf(updateApiEntity.getLifecycleState().name()));
-        }
-        return api;
-    }
-
     private LifecycleState convert(EventType eventType) {
         LifecycleState lifecycleState;
         switch (eventType) {
@@ -3502,72 +1748,6 @@ public class ApiServiceImpl extends AbstractService implements ApiService, ApiDu
                 throw new IllegalArgumentException("Unknown EventType " + eventType.toString() + " to convert EventType into Lifecycle");
         }
         return lifecycleState;
-    }
-
-    private static class MemberToImport {
-
-        private String source;
-        private String sourceId;
-        private List<String> roles; // After v3
-        private String role; // Before v3
-
-        public MemberToImport() {}
-
-        public MemberToImport(String source, String sourceId, List<String> roles, String role) {
-            this.source = source;
-            this.sourceId = sourceId;
-            this.roles = roles;
-            this.role = role;
-        }
-
-        public String getSource() {
-            return source;
-        }
-
-        public void setSource(String source) {
-            this.source = source;
-        }
-
-        public String getSourceId() {
-            return sourceId;
-        }
-
-        public void setSourceId(String sourceId) {
-            this.sourceId = sourceId;
-        }
-
-        public List<String> getRoles() {
-            return roles;
-        }
-
-        public void setRoles(List<String> roles) {
-            this.roles = roles;
-        }
-
-        public String getRole() {
-            return role;
-        }
-
-        public void setRole(String role) {
-            this.role = role;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            MemberToImport that = (MemberToImport) o;
-
-            return source.equals(that.source) && sourceId.equals(that.sourceId);
-        }
-
-        @Override
-        public int hashCode() {
-            int result = source.hashCode();
-            result = 31 * result + sourceId.hashCode();
-            return result;
-        }
     }
 
     private ProxyModelEntity convert(Proxy proxy) {
