+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/ApiDuplicatorService_CreateWithDefinitionTest.java
@@ -32,6 +32,7 @@ import io.gravitee.rest.api.model.parameters.ParameterReferenceType;
 import io.gravitee.rest.api.model.permissions.RoleScope;
 import io.gravitee.rest.api.model.permissions.SystemRole;
 import io.gravitee.rest.api.service.exceptions.PlanNotFoundException;
+import io.gravitee.rest.api.service.impl.ApiDuplicatorServiceImpl;
 import io.gravitee.rest.api.service.impl.ApiServiceImpl;
 import io.gravitee.rest.api.service.notification.NotificationTemplateService;
 import io.gravitee.rest.api.service.search.SearchEngineService;
@@ -41,7 +42,6 @@ import java.io.Reader;
 import java.net.URL;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.Optional;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -61,14 +61,14 @@ import org.springframework.security.core.context.SecurityContextHolder;
  * @author GraviteeSource Team
  */
 @RunWith(MockitoJUnitRunner.class)
-public class ApiService_CreateWithDefinitionTest {
+public class ApiDuplicatorService_CreateWithDefinitionTest {
 
     private static final String API_ID = "id-api";
     private static final String PLAN_ID = "my-plan";
     private static final String SOURCE = "source";
 
     @InjectMocks
-    private ApiServiceImpl apiService = new ApiServiceImpl();
+    protected ApiDuplicatorService apiDuplicatorService = new ApiDuplicatorServiceImpl();
 
     @Mock
     private ApiRepository apiRepository;
@@ -202,7 +202,7 @@ public class ApiService_CreateWithDefinitionTest {
         when(userService.findById(memberEntity.getId())).thenReturn(admin);
         when(pageService.createWithDefinition(any(), any())).thenReturn(new PageEntity());
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(2)).createWithDefinition(eq(API_ID), anyString());
         verify(membershipService, times(1))
@@ -264,7 +264,7 @@ public class ApiService_CreateWithDefinitionTest {
         memberEntity.setRoles(Collections.singletonList(poRoleEntity));
         when(membershipService.addRoleToMemberOnReference(any(), any(), any())).thenReturn(memberEntity);
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(1)).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, times(1))
@@ -303,7 +303,7 @@ public class ApiService_CreateWithDefinitionTest {
         when(userService.findById(admin.getId())).thenReturn(admin);
         when(pageService.createWithDefinition(any(), any())).thenReturn(new PageEntity());
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(2)).createWithDefinition(eq(API_ID), anyString());
         verify(apiRepository, never()).update(any());
@@ -332,7 +332,7 @@ public class ApiService_CreateWithDefinitionTest {
         user.setSourceId(API_ID);
         when(userService.findById(admin.getId())).thenReturn(admin);
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
         verify(apiRepository, never()).update(any());
@@ -369,7 +369,7 @@ public class ApiService_CreateWithDefinitionTest {
         when(userService.findById(any())).thenReturn(user);
         when(parameterService.find(Key.API_PRIMARY_OWNER_MODE, ParameterReferenceType.ENVIRONMENT)).thenReturn("HYBRID");
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
         verify(apiRepository, never()).update(any());
@@ -411,7 +411,7 @@ public class ApiService_CreateWithDefinitionTest {
         when(groupService.findById(any())).thenReturn(group);
         when(parameterService.find(Key.API_PRIMARY_OWNER_MODE, ParameterReferenceType.ENVIRONMENT)).thenReturn("GROUP");
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
         verify(apiRepository, never()).update(any());
@@ -447,7 +447,7 @@ public class ApiService_CreateWithDefinitionTest {
         user.setSourceId(API_ID);
         when(userService.findById(admin.getId())).thenReturn(admin);
 
-        ApiEntity apiEntityCreated = apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        ApiEntity apiEntityCreated = apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
         verify(apiRepository, never()).update(any());
@@ -473,7 +473,7 @@ public class ApiService_CreateWithDefinitionTest {
         when(userService.findById(admin.getId())).thenReturn(admin);
         when(planService.findById(anyString())).thenThrow(PlanNotFoundException.class);
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(pageService, times(1)).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, times(1))
@@ -508,7 +508,7 @@ public class ApiService_CreateWithDefinitionTest {
         when(userService.findById(admin.getId())).thenReturn(admin);
         when(planService.findById(anyString())).thenThrow(PlanNotFoundException.class);
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(planService, times(2)).create(any(NewPlanEntity.class));
         verify(apiRepository, never()).update(any());
@@ -533,7 +533,7 @@ public class ApiService_CreateWithDefinitionTest {
         admin.setSourceId(API_ID);
         when(userService.findById(admin.getId())).thenReturn(admin);
 
-        apiService.createWithImportedDefinition(null, toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(null, toBeImport, "admin");
 
         verify(apiMetadataService, times(1)).create(any(NewApiMetadataEntity.class));
         verify(apiMetadataService, times(2)).update(any(UpdateApiMetadataEntity.class));
