+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/ApiDuplicatorService_ExportAsJsonTestSetup.java
@@ -30,7 +30,6 @@ import com.google.common.io.Resources;
 import io.gravitee.common.http.HttpMethod;
 import io.gravitee.definition.jackson.datatype.GraviteeMapper;
 import io.gravitee.definition.model.*;
-import io.gravitee.definition.model.endpoint.GrpcEndpoint;
 import io.gravitee.definition.model.endpoint.HttpEndpoint;
 import io.gravitee.repository.exceptions.TechnicalException;
 import io.gravitee.repository.management.api.ApiRepository;
@@ -42,7 +41,7 @@ import io.gravitee.rest.api.model.parameters.Key;
 import io.gravitee.rest.api.model.parameters.ParameterReferenceType;
 import io.gravitee.rest.api.model.permissions.SystemRole;
 import io.gravitee.rest.api.service.common.GraviteeContext;
-import io.gravitee.rest.api.service.impl.ApiServiceImpl;
+import io.gravitee.rest.api.service.impl.ApiDuplicatorServiceImpl;
 import io.gravitee.rest.api.service.jackson.filter.ApiPermissionFilter;
 import io.gravitee.rest.api.service.jackson.ser.api.*;
 import java.io.IOException;
@@ -57,26 +56,25 @@ import org.mockito.internal.util.collections.Sets;
 import org.springframework.context.ApplicationContext;
 
 /**
- * @author Eric LELEU (eric.leleu at graviteesource.com)
  * @author Azize Elamrani (azize.elamrani at graviteesource.com)
  * @author Nicolas Geraud (nicolas.geraud at graviteesource.com)
  * @author GraviteeSource Team
  */
-public class ApiService_gRPC_ExportAsJsonTestSetup {
+public class ApiDuplicatorService_ExportAsJsonTestSetup {
 
-    private static final String API_ID = "id-api";
+    protected static final String API_ID = "id-api";
 
     @InjectMocks
-    private ApiServiceImpl apiService = new ApiServiceImpl();
+    protected ApiDuplicatorService apiDuplicatorService = new ApiDuplicatorServiceImpl();
 
     @Mock
-    private ApiRepository apiRepository;
+    protected ApiRepository apiRepository;
 
     @Mock
     private MembershipRepository membershipRepository;
 
     @Spy
-    private ObjectMapper objectMapper = new GraviteeMapper();
+    protected ObjectMapper objectMapper = new GraviteeMapper();
 
     @Mock
     private MembershipService membershipService;
@@ -107,6 +105,7 @@ public class ApiService_gRPC_ExportAsJsonTestSetup {
 
     @Before
     public void setUp() throws TechnicalException {
+        GraviteeContext.setCurrentEnvironment("DEFAULT");
         PropertyFilter apiMembershipTypeFilter = new ApiPermissionFilter();
         objectMapper.setFilterProvider(
             new SimpleFilterProvider(Collections.singletonMap("apiMembershipTypeFilter", apiMembershipTypeFilter))
@@ -158,9 +157,9 @@ public class ApiService_gRPC_ExportAsJsonTestSetup {
 
         Api api = new Api();
         api.setId(API_ID);
-        String definition = null;
         api.setEnvironmentId("DEFAULT");
 
+        String definition = null;
         try {
             definition = objectMapper.writeValueAsString(buildApiDefinition(api));
         } catch (JsonProcessingException e) {}
@@ -256,7 +255,7 @@ public class ApiService_gRPC_ExportAsJsonTestSetup {
         publishedPlan.setValidation(PlanValidationType.AUTO);
         publishedPlan.setStatus(PlanStatus.PUBLISHED);
         Map<String, List<Rule>> paths = new HashMap<>();
-        Rule rule = new Rule();
+        io.gravitee.definition.model.Rule rule = new io.gravitee.definition.model.Rule();
         rule.setEnabled(true);
         rule.setMethods(Sets.newSet(HttpMethod.GET));
         Policy policy = new Policy();
@@ -310,6 +309,7 @@ public class ApiService_gRPC_ExportAsJsonTestSetup {
 
     protected io.gravitee.definition.model.Api buildApiDefinition(Api api) {
         api.setDescription("Gravitee.io");
+
         // set proxy
         Proxy proxy = new Proxy();
         proxy.setVirtualHosts(Collections.singletonList(new VirtualHost("/test")));
@@ -321,18 +321,16 @@ public class ApiService_gRPC_ExportAsJsonTestSetup {
         EndpointGroup endpointGroup = new EndpointGroup();
         endpointGroup.setName("default-group");
         Endpoint endpoint = new HttpEndpoint("default", "http://test");
-        Endpoint endPointGrpc = new GrpcEndpoint("EndPoint GRPC", "grpc://localhost:8888");
-        LinkedHashSet endpoints = new LinkedHashSet();
-        endpoints.add(endpoint);
-        endpoints.add(endPointGrpc);
-        endpointGroup.setEndpoints(endpoints);
+        endpointGroup.setEndpoints(Collections.singleton(endpoint));
         LoadBalancer loadBalancer = new LoadBalancer();
         loadBalancer.setType(LoadBalancerType.ROUND_ROBIN);
         endpointGroup.setLoadBalancer(loadBalancer);
         proxy.setGroups(Collections.singleton(endpointGroup));
+
         io.gravitee.definition.model.Api apiDefinition = new io.gravitee.definition.model.Api();
-        apiDefinition.setProxy(proxy);
         apiDefinition.setPaths(Collections.emptyMap());
+        apiDefinition.setProxy(proxy);
+
         ResponseTemplate responseTemplate = new ResponseTemplate();
         responseTemplate.setStatusCode(400);
         responseTemplate.setBody("{\"bad\":\"news\"}");
@@ -340,11 +338,65 @@ public class ApiService_gRPC_ExportAsJsonTestSetup {
         return apiDefinition;
     }
 
-    protected void shouldConvertAsJsonForExport(ApiSerializer.Version version, String filename) throws TechnicalException, IOException {
-        String jsonForExport = apiService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name());
+    protected void shouldConvertAsJsonForExport(ApiSerializer.Version version, String filename) throws IOException {
+        String jsonForExport = apiDuplicatorService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name());
 
         URL url = Resources.getResource(
-            "io/gravitee/rest/api/management/service/export-gRPC-convertAsJsonForExport" +
+            "io/gravitee/rest/api/management/service/export-convertAsJsonForExport" + (filename != null ? "-" + filename : "") + ".json"
+        );
+        String expectedJson = Resources.toString(url, Charsets.UTF_8);
+
+        assertThat(jsonForExport).isNotNull();
+        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
+    }
+
+    protected void shouldConvertAsJsonWithoutMembers(ApiSerializer.Version version, String filename) throws IOException {
+        String jsonForExport = apiDuplicatorService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "members");
+
+        URL url = Resources.getResource(
+            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutMembers" +
+            (filename != null ? "-" + filename : "") +
+            ".json"
+        );
+        String expectedJson = Resources.toString(url, Charsets.UTF_8);
+
+        assertThat(jsonForExport).isNotNull();
+        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
+    }
+
+    protected void shouldConvertAsJsonWithoutPages(ApiSerializer.Version version, String filename) throws IOException {
+        String jsonForExport = apiDuplicatorService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "pages");
+
+        URL url = Resources.getResource(
+            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutPages" +
+            (filename != null ? "-" + filename : "") +
+            ".json"
+        );
+        String expectedJson = Resources.toString(url, Charsets.UTF_8);
+
+        assertThat(jsonForExport).isNotNull();
+        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
+    }
+
+    protected void shouldConvertAsJsonWithoutPlans(ApiSerializer.Version version, String filename) throws IOException {
+        String jsonForExport = apiDuplicatorService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "plans");
+
+        URL url = Resources.getResource(
+            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutPlans" +
+            (filename != null ? "-" + filename : "") +
+            ".json"
+        );
+        String expectedJson = Resources.toString(url, Charsets.UTF_8);
+
+        assertThat(jsonForExport).isNotNull();
+        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
+    }
+
+    protected void shouldConvertAsJsonWithoutMetadata(ApiSerializer.Version version, String filename) throws IOException {
+        String jsonForExport = apiDuplicatorService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "metadata");
+
+        URL url = Resources.getResource(
+            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutMetadata" +
             (filename != null ? "-" + filename : "") +
             ".json"
         );
