+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/ApiDuplicatorService_UpdateWithDefinitionTest.java
@@ -31,6 +31,7 @@ import io.gravitee.rest.api.model.*;
 import io.gravitee.rest.api.model.api.ApiEntity;
 import io.gravitee.rest.api.model.permissions.RoleScope;
 import io.gravitee.rest.api.model.permissions.SystemRole;
+import io.gravitee.rest.api.service.impl.ApiDuplicatorServiceImpl;
 import io.gravitee.rest.api.service.impl.ApiServiceImpl;
 import io.gravitee.rest.api.service.notification.NotificationTemplateService;
 import io.gravitee.rest.api.service.search.SearchEngineService;
@@ -59,14 +60,14 @@ import org.springframework.security.core.context.SecurityContextHolder;
  * @author GraviteeSource Team
  */
 @RunWith(MockitoJUnitRunner.class)
-public class ApiService_UpdateWithDefinitionTest {
+public class ApiDuplicatorService_UpdateWithDefinitionTest {
 
     private static final String API_ID = "id-api";
     private static final String PLAN_ID = "my-plan";
     private static final String SOURCE = "source";
 
     @InjectMocks
-    private ApiServiceImpl apiService = new ApiServiceImpl();
+    protected ApiDuplicatorService apiDuplicatorService = new ApiDuplicatorServiceImpl();
 
     @Mock
     private ApiRepository apiRepository;
@@ -200,7 +201,7 @@ public class ApiService_UpdateWithDefinitionTest {
         when(membershipService.addRoleToMemberOnReference(any(), any(), any(), any(), any())).thenReturn(memberEntity);
         when(userService.findBySource(user.getSource(), user.getSourceId(), false)).thenReturn(user);
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
 
         verify(pageService, times(2)).update(any(), any(UpdatePageEntity.class));
         verify(membershipService, never())
@@ -286,7 +287,7 @@ public class ApiService_UpdateWithDefinitionTest {
             .thenReturn(new HashSet(Arrays.asList(po)));
         when(userService.findById(admin.getId())).thenReturn(admin);
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
 
         verify(pageService, never()).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, never())
@@ -315,7 +316,7 @@ public class ApiService_UpdateWithDefinitionTest {
         when(userService.findById(admin.getId())).thenReturn(admin);
         when(userService.findById(user.getId())).thenReturn(user);
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
 
         verify(pageService, never()).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, never())
@@ -351,7 +352,7 @@ public class ApiService_UpdateWithDefinitionTest {
         when(userService.findById(admin.getId())).thenReturn(admin);
         when(userService.findById(user.getId())).thenReturn(user);
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
 
         verify(pageService, never()).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, never())
@@ -403,7 +404,7 @@ public class ApiService_UpdateWithDefinitionTest {
 
         PageEntity existingPage = mock(PageEntity.class);
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, null);
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, null);
 
         verify(pageService, times(2)).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, never()).addRoleToMemberOnReference(any(), any(), any());
@@ -441,7 +442,7 @@ public class ApiService_UpdateWithDefinitionTest {
         )
             .thenReturn(Collections.singleton(po));
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, null);
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, null);
 
         verify(pageService, never()).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, never()).addRoleToMemberOnReference(any(), any(), any());
@@ -464,7 +465,7 @@ public class ApiService_UpdateWithDefinitionTest {
         plan2.setId("plan-id2");
         when(planService.findById(plan2.getId())).thenReturn(plan2);
 
-        apiService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
+        apiDuplicatorService.updateWithImportedDefinition(apiEntity, toBeImport, "import");
 
         verify(planService, times(2)).update(any(UpdatePlanEntity.class));
         verify(apiRepository, times(1)).update(any());
