+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/ApiDuplicatorService_gRPC_ExportAsJsonTestSetup.java
@@ -30,6 +30,7 @@ import com.google.common.io.Resources;
 import io.gravitee.common.http.HttpMethod;
 import io.gravitee.definition.jackson.datatype.GraviteeMapper;
 import io.gravitee.definition.model.*;
+import io.gravitee.definition.model.endpoint.GrpcEndpoint;
 import io.gravitee.definition.model.endpoint.HttpEndpoint;
 import io.gravitee.repository.exceptions.TechnicalException;
 import io.gravitee.repository.management.api.ApiRepository;
@@ -41,6 +42,7 @@ import io.gravitee.rest.api.model.parameters.Key;
 import io.gravitee.rest.api.model.parameters.ParameterReferenceType;
 import io.gravitee.rest.api.model.permissions.SystemRole;
 import io.gravitee.rest.api.service.common.GraviteeContext;
+import io.gravitee.rest.api.service.impl.ApiDuplicatorServiceImpl;
 import io.gravitee.rest.api.service.impl.ApiServiceImpl;
 import io.gravitee.rest.api.service.jackson.filter.ApiPermissionFilter;
 import io.gravitee.rest.api.service.jackson.ser.api.*;
@@ -56,25 +58,26 @@ import org.mockito.internal.util.collections.Sets;
 import org.springframework.context.ApplicationContext;
 
 /**
+ * @author Eric LELEU (eric.leleu at graviteesource.com)
  * @author Azize Elamrani (azize.elamrani at graviteesource.com)
  * @author Nicolas Geraud (nicolas.geraud at graviteesource.com)
  * @author GraviteeSource Team
  */
-public class ApiService_ExportAsJsonTestSetup {
+public class ApiDuplicatorService_gRPC_ExportAsJsonTestSetup {
 
-    protected static final String API_ID = "id-api";
+    private static final String API_ID = "id-api";
 
     @InjectMocks
-    protected ApiServiceImpl apiService = new ApiServiceImpl();
+    protected ApiDuplicatorService apiDuplicatorService = new ApiDuplicatorServiceImpl();
 
     @Mock
-    protected ApiRepository apiRepository;
+    private ApiRepository apiRepository;
 
     @Mock
     private MembershipRepository membershipRepository;
 
     @Spy
-    protected ObjectMapper objectMapper = new GraviteeMapper();
+    private ObjectMapper objectMapper = new GraviteeMapper();
 
     @Mock
     private MembershipService membershipService;
@@ -105,7 +108,6 @@ public class ApiService_ExportAsJsonTestSetup {
 
     @Before
     public void setUp() throws TechnicalException {
-        GraviteeContext.setCurrentEnvironment("DEFAULT");
         PropertyFilter apiMembershipTypeFilter = new ApiPermissionFilter();
         objectMapper.setFilterProvider(
             new SimpleFilterProvider(Collections.singletonMap("apiMembershipTypeFilter", apiMembershipTypeFilter))
@@ -157,9 +159,9 @@ public class ApiService_ExportAsJsonTestSetup {
 
         Api api = new Api();
         api.setId(API_ID);
+        String definition = null;
         api.setEnvironmentId("DEFAULT");
 
-        String definition = null;
         try {
             definition = objectMapper.writeValueAsString(buildApiDefinition(api));
         } catch (JsonProcessingException e) {}
@@ -255,7 +257,7 @@ public class ApiService_ExportAsJsonTestSetup {
         publishedPlan.setValidation(PlanValidationType.AUTO);
         publishedPlan.setStatus(PlanStatus.PUBLISHED);
         Map<String, List<Rule>> paths = new HashMap<>();
-        io.gravitee.definition.model.Rule rule = new io.gravitee.definition.model.Rule();
+        Rule rule = new Rule();
         rule.setEnabled(true);
         rule.setMethods(Sets.newSet(HttpMethod.GET));
         Policy policy = new Policy();
@@ -309,7 +311,6 @@ public class ApiService_ExportAsJsonTestSetup {
 
     protected io.gravitee.definition.model.Api buildApiDefinition(Api api) {
         api.setDescription("Gravitee.io");
-
         // set proxy
         Proxy proxy = new Proxy();
         proxy.setVirtualHosts(Collections.singletonList(new VirtualHost("/test")));
@@ -321,16 +322,18 @@ public class ApiService_ExportAsJsonTestSetup {
         EndpointGroup endpointGroup = new EndpointGroup();
         endpointGroup.setName("default-group");
         Endpoint endpoint = new HttpEndpoint("default", "http://test");
-        endpointGroup.setEndpoints(Collections.singleton(endpoint));
+        Endpoint endPointGrpc = new GrpcEndpoint("EndPoint GRPC", "grpc://localhost:8888");
+        LinkedHashSet endpoints = new LinkedHashSet();
+        endpoints.add(endpoint);
+        endpoints.add(endPointGrpc);
+        endpointGroup.setEndpoints(endpoints);
         LoadBalancer loadBalancer = new LoadBalancer();
         loadBalancer.setType(LoadBalancerType.ROUND_ROBIN);
         endpointGroup.setLoadBalancer(loadBalancer);
         proxy.setGroups(Collections.singleton(endpointGroup));
-
         io.gravitee.definition.model.Api apiDefinition = new io.gravitee.definition.model.Api();
-        apiDefinition.setPaths(Collections.emptyMap());
         apiDefinition.setProxy(proxy);
-
+        apiDefinition.setPaths(Collections.emptyMap());
         ResponseTemplate responseTemplate = new ResponseTemplate();
         responseTemplate.setStatusCode(400);
         responseTemplate.setBody("{\"bad\":\"news\"}");
@@ -338,65 +341,11 @@ public class ApiService_ExportAsJsonTestSetup {
         return apiDefinition;
     }
 
-    protected void shouldConvertAsJsonForExport(ApiSerializer.Version version, String filename) throws IOException {
-        String jsonForExport = apiService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name());
+    protected void shouldConvertAsJsonForExport(ApiSerializer.Version version, String filename) throws TechnicalException, IOException {
+        String jsonForExport = apiDuplicatorService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name());
 
         URL url = Resources.getResource(
-            "io/gravitee/rest/api/management/service/export-convertAsJsonForExport" + (filename != null ? "-" + filename : "") + ".json"
-        );
-        String expectedJson = Resources.toString(url, Charsets.UTF_8);
-
-        assertThat(jsonForExport).isNotNull();
-        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
-    }
-
-    protected void shouldConvertAsJsonWithoutMembers(ApiSerializer.Version version, String filename) throws IOException {
-        String jsonForExport = apiService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "members");
-
-        URL url = Resources.getResource(
-            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutMembers" +
-            (filename != null ? "-" + filename : "") +
-            ".json"
-        );
-        String expectedJson = Resources.toString(url, Charsets.UTF_8);
-
-        assertThat(jsonForExport).isNotNull();
-        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
-    }
-
-    protected void shouldConvertAsJsonWithoutPages(ApiSerializer.Version version, String filename) throws IOException {
-        String jsonForExport = apiService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "pages");
-
-        URL url = Resources.getResource(
-            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutPages" +
-            (filename != null ? "-" + filename : "") +
-            ".json"
-        );
-        String expectedJson = Resources.toString(url, Charsets.UTF_8);
-
-        assertThat(jsonForExport).isNotNull();
-        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
-    }
-
-    protected void shouldConvertAsJsonWithoutPlans(ApiSerializer.Version version, String filename) throws IOException {
-        String jsonForExport = apiService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "plans");
-
-        URL url = Resources.getResource(
-            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutPlans" +
-            (filename != null ? "-" + filename : "") +
-            ".json"
-        );
-        String expectedJson = Resources.toString(url, Charsets.UTF_8);
-
-        assertThat(jsonForExport).isNotNull();
-        assertThat(objectMapper.readTree(jsonForExport)).isEqualTo(objectMapper.readTree(expectedJson));
-    }
-
-    protected void shouldConvertAsJsonWithoutMetadata(ApiSerializer.Version version, String filename) throws IOException {
-        String jsonForExport = apiService.exportAsJson(API_ID, version.getVersion(), SystemRole.PRIMARY_OWNER.name(), "metadata");
-
-        URL url = Resources.getResource(
-            "io/gravitee/rest/api/management/service/export-convertAsJsonForExportWithoutMetadata" +
+            "io/gravitee/rest/api/management/service/export-gRPC-convertAsJsonForExport" +
             (filename != null ? "-" + filename : "") +
             ".json"
         );
